build: add more logos
