+++ src/components/logos/Logos.stories.js
@@ -1,768 +1,9 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
-import { Flex, Svg, Subhead } from 'components/elements'
-import { createCssState } from 'helpers/style'
-import styled, { css } from 'styled-components'
-import { transition } from 'theme'
+import { Flex, Subhead } from 'components/elements'
+import * as Logos from 'components/logos'
 
-const style = css`
-  .logo {
-    transition: fill ${transition.medium}, stroke ${transition.medium};
-  }
-`
-
-const FacebookBase = props => (
-  <Svg
-    version='1.1'
-    height={41.5}
-    viewBox='0 0 1024.0159 205.05107'
-    overflow='visible'
-    enableBackground='new 0 0 266 100'
-    xmlSpace='preserve'
-    {...props}
-  >
-    <path
-      className='logo'
-      d='M533.69 99.649c-8.561 0-14.732 2.806-20.987 5.667v64.706c5.993.571 9.427.571 15.115.571 20.547 0 23.363-9.408 23.363-22.547V117.14c0-9.7-3.22-17.491-17.492-17.491zm-136.447-3.531c-14.254 0-17.51 7.827-17.51 17.517v5.445h34.999v-5.445c0-9.69-3.252-17.517-17.49-17.517zm-264.324 67.353c0 7.666 3.617 11.648 11.601 11.648 8.565 0 13.635-2.792 19.886-5.66v-15.35h-18.728c-8.863 0-12.76 1.65-12.76 9.362zm533.5-63.822c-14.278 0-19.226 7.792-19.226 17.49v35.405c0 9.728 4.948 17.543 19.225 17.543 14.24 0 19.225-7.815 19.225-17.543V117.14c0-9.7-4.985-17.491-19.225-17.491zM62.945 203.167h-41.97V101.572H0V66.564h20.975v-21.02C20.975 16.983 32.813 0 66.455 0h28.011v35.015H76.96c-13.099 0-13.961 4.892-13.961 14.024l-.052 17.525h31.717l-3.71 35.008H62.945v101.595zm143.447.263h-34.98l-1.516-8.847c-15.973 8.847-30.229 10.28-39.633 10.28-25.653 0-39.312-17.14-39.312-40.839 0-27.964 15.933-37.944 44.44-37.944h29.014v-6.044c0-14.27-1.635-18.464-23.585-18.464h-35.89l3.51-35.008h39.23c48.164 0 58.722 15.214 58.722 53.753v83.113zm118.958-99.29c-21.769-3.73-28.02-4.55-38.494-4.55-18.815 0-24.501 4.151-24.501 20.127v30.222c0 15.978 5.686 20.148 24.5 20.148 10.476 0 16.726-.834 38.495-4.582v34.148c-19.068 4.274-31.49 5.398-41.99 5.398-45.063 0-62.976-23.7-62.976-57.933v-24.524c0-34.263 17.913-58.003 62.976-58.003 10.5 0 22.922 1.13 41.99 5.421v34.128zm131.352 42.978h-76.97v2.821c0 15.978 5.69 20.148 24.5 20.148 16.91 0 27.23-.834 48.96-4.582v34.148c-20.956 4.274-31.876 5.398-52.439 5.398-45.066 0-62.988-23.7-62.988-57.933V119.08c0-29.953 13.298-54.49 59.477-54.49 46.178 0 59.46 24.251 59.46 54.49v28.038zm136.442.647c0 33.09-9.456 57.22-66.746 57.22-20.687 0-32.823-1.818-55.656-5.332V10.513l41.96-7.003v66.16c9.066-3.369 20.804-5.08 31.487-5.08 41.964 0 48.955 18.814 48.955 49.045v34.13zm134.505.722c0 28.544-11.785 56.227-61.091 56.227-49.33 0-61.335-27.683-61.335-56.227v-27.561c0-28.559 12.005-56.246 61.335-56.246 49.306 0 61.09 27.687 61.09 56.246v27.56zm134.41 0c0 28.544-11.798 56.227-61.095 56.227-49.33 0-61.334-27.683-61.334-56.227v-27.561c0-28.559 12.004-56.246 61.334-56.246 49.297 0 61.096 27.687 61.096 56.246v27.56zm137.94 54.68h-45.483l-38.464-64.211v64.211h-41.974V10.512l41.974-7.002v124.023l38.464-60.97H1000L958.008 133.1 1000 203.167zM800.825 99.65c-14.254 0-19.202 7.792-19.202 17.49v35.405c0 9.728 4.948 17.543 19.202 17.543 14.235 0 19.272-7.815 19.272-17.543V117.14c0-9.7-5.037-17.491-19.272-17.491zm213.241 83.65c5.537 0 9.95 4.516 9.95 10.156 0 5.727-4.413 10.192-9.99 10.192-5.547 0-10.049-4.465-10.049-10.192 0-5.64 4.502-10.155 10.049-10.155h.04zm-.04 1.58c-4.46 0-8.113 3.84-8.113 8.576 0 4.824 3.653 8.613 8.154 8.613 4.505.044 8.103-3.79 8.103-8.57s-3.598-8.62-8.103-8.62h-.04zm-1.894 14.487h-1.806v-11.33c.947-.132 1.85-.265 3.201-.265 1.715 0 2.834.36 3.52.852.666.5 1.025 1.263 1.025 2.342 0 1.498-.984 2.394-2.199 2.761v.088c.988.184 1.663 1.08 1.89 2.747.265 1.762.537 2.438.717 2.805h-1.891c-.268-.368-.54-1.403-.767-2.893-.265-1.44-.992-1.983-2.438-1.983h-1.252v4.876zm0-6.271h1.307c1.48 0 2.739-.543 2.739-1.946 0-.991-.716-1.983-2.74-1.983-.59 0-.998.045-1.306.089v3.84z'
-    />
-  </Svg>
-)
-
-const Facebook = styled(FacebookBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #3b5998;
-      }
-    `
-  })};
-`
-
-const GitHubBase = props => (
-  <Svg
-    height={73.5}
-    viewBox='0 0 2880 1024'
-    fill='none'
-    xmlns='http://www.w3.org/2000/svg'
-    {...props}
-  >
-    <path
-      className='logo'
-      fillRule='evenodd'
-      clipRule='evenodd'
-      d='M18.53 10.0301H18.51C18.519 10.0301 18.525 10.0391 18.534 10.0411C18.535 10.0411 18.539 10.0401 18.54 10.0401L18.53 10.0301ZM18.534 10.0411C18.441 10.0421 18.207 10.0901 17.96 10.0901C17.18 10.0901 16.91 9.73007 16.91 9.26007V6.13007H18.5C18.59 6.13007 18.66 6.05007 18.66 5.94007V4.24007C18.66 4.15007 18.58 4.07007 18.5 4.07007H16.91V1.96007C16.91 1.88007 16.86 1.83007 16.77 1.83007H14.61C14.52 1.83007 14.47 1.88007 14.47 1.96007V4.13007C14.47 4.13007 13.38 4.40007 13.31 4.41007C13.23 4.43007 13.18 4.50007 13.18 4.58007V5.94007C13.18 6.05007 13.26 6.13007 13.35 6.13007H14.46V9.41007C14.46 11.8501 16.16 12.1001 17.32 12.1001C17.85 12.1001 18.49 11.9301 18.59 11.8801C18.65 11.8601 18.68 11.7901 18.68 11.7201V10.2201C18.68 10.1211 18.612 10.0551 18.534 10.0411ZM42.23 7.84007C42.23 6.03007 41.5 5.79007 40.73 5.87007C40.13 5.91007 39.65 6.21007 39.65 6.21007V9.73007C39.65 9.73007 40.14 10.0701 40.87 10.0901C41.9 10.1201 42.23 9.75007 42.23 7.84007ZM44.66 7.68007C44.66 11.1101 43.55 12.0901 41.61 12.0901C39.97 12.0901 39.09 11.2601 39.09 11.2601C39.09 11.2601 39.05 11.7201 39 11.7801C38.97 11.8401 38.92 11.8601 38.86 11.8601H37.38C37.28 11.8601 37.19 11.7801 37.19 11.6901L37.21 0.580069C37.21 0.490068 37.29 0.410069 37.38 0.410069H39.51C39.6 0.410069 39.68 0.490068 39.68 0.580069V4.35007C39.68 4.35007 40.5 3.82007 41.7 3.82007L41.69 3.80007C42.89 3.80007 44.66 4.25007 44.66 7.68007ZM35.94 4.07007H35.93H33.84C33.73 4.07007 33.67 4.15007 33.67 4.26007V9.70007C33.67 9.70007 33.12 10.0901 32.37 10.0901C31.62 10.0901 31.4 9.75007 31.4 9.00007V4.25007C31.4 4.16007 31.32 4.08007 31.23 4.08007H29.09C29 4.08007 28.92 4.16007 28.92 4.25007V9.36007C28.92 11.5601 30.15 12.1101 31.84 12.1101C33.23 12.1101 34.36 11.3401 34.36 11.3401C34.36 11.3401 34.41 11.7301 34.44 11.7901C34.46 11.8401 34.53 11.8801 34.6 11.8801H35.94C36.05 11.8801 36.11 11.8001 36.11 11.7101L36.13 4.24007C36.13 4.15007 36.05 4.07007 35.94 4.07007ZM12.24 4.06007H10.11C10.02 4.06007 9.94 4.15007 9.94 4.26007V11.6001C9.94 11.8001 10.07 11.8701 10.24 11.8701H12.16C12.36 11.8701 12.41 11.7801 12.41 11.6001V4.24007V4.23007C12.41 4.14007 12.33 4.06007 12.24 4.06007ZM11.19 0.680069C10.42 0.680069 9.81 1.29007 9.81 2.06007C9.81 2.83007 10.42 3.44007 11.19 3.44007C11.94 3.44007 12.55 2.83007 12.55 2.06007C12.55 1.29007 11.94 0.680069 11.19 0.680069ZM27.68 0.430069H25.57C25.48 0.430069 25.4 0.510069 25.4 0.600069V4.69007H22.09V0.600069C22.09 0.510069 22.01 0.430069 21.92 0.430069H19.79C19.7 0.430069 19.62 0.510069 19.62 0.600069V11.7101C19.62 11.8001 19.71 11.8801 19.79 11.8801H21.92C22.01 11.8801 22.09 11.8001 22.09 11.7101V6.96007H25.4L25.38 11.7101C25.38 11.8001 25.46 11.8801 25.55 11.8801H27.68C27.77 11.8801 27.85 11.8001 27.85 11.7101V0.600069C27.85 0.510069 27.77 0.430069 27.68 0.430069ZM8.81 5.35007V11.0901C8.81 11.1301 8.8 11.2001 8.75 11.2201C8.75 11.2201 7.5 12.1101 5.44 12.1101C2.95 12.1101 0 11.3301 0 6.19007C0 1.05007 2.58 -0.00993021 5.1 6.97893e-05C7.28 6.97893e-05 8.16 0.49007 8.3 0.58007C8.34 0.63007 8.36 0.67007 8.36 0.72007L7.94 2.50007C7.94 2.59007 7.85 2.70007 7.74 2.67007C7.38 2.56007 6.84 2.34007 5.57 2.34007C4.1 2.34007 2.52 2.76007 2.52 6.07007C2.52 9.38007 4.02 9.77007 5.1 9.77007C6.02 9.77007 6.35 9.66007 6.35 9.66007V7.36007H4.88C4.77 7.36007 4.69 7.28007 4.69 7.19007V5.35007C4.69 5.26007 4.77 5.18007 4.88 5.18007H8.62C8.73 5.18007 8.81 5.26007 8.81 5.35007Z'
-      transform='matrix(64 0 0 64 0 128.002)'
-      fill='black'
-    />
-  </Svg>
-)
-
-const GitHub = styled(GitHubBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #151013;
-      }
-    `
-  })};
-`
-
-const MediumBase = props => (
-  <Svg
-    height={40.8}
-    viewBox='0 0 380 75'
-    version='1.1'
-    xmlnsXlink='http://www.w3.org/1999/xlink'
-    {...props}
-  >
-    <path
-      className='logo'
-      d='M191.928 66.629c0 1.32 0 1.62.89 2.43l4.985 4.92v.21H175.7v-9.45c-2.225 6.228-8.109 10.347-14.657 10.26-11.363 0-19.611-9-19.611-25.56 0-15.51 8.9-25.86 21.391-25.86 5.537-.352 10.69 2.873 12.847 8.04V10.77a2.594 2.594 0 0 0-1.009-2.52l-4.747-4.62v-.21L191.928.81v65.819zm-16.228-3v-31.11a8.242 8.242 0 0 0-6.587-3.12c-5.37 0-10.562 4.92-10.562 19.53 0 13.68 4.48 18.51 10.058 18.51a8.874 8.874 0 0 0 7.09-3.81zm30.885 10.56v-42a2.838 2.838 0 0 0-.89-2.52l-4.895-5.1v-.21h22.103v42.6c0 1.32 0 1.62.89 2.43l4.866 4.62v.21l-22.074-.03zm-.801-65.19c0-4.97 3.985-8.999 8.9-8.999 4.916 0 8.901 4.03 8.901 9s-3.985 9-8.9 9c-4.916 0-8.901-4.03-8.901-9zm76.753 57.54c0 1.32 0 1.71.89 2.52l4.985 4.92v.21H266.19v-10.77c-2.024 6.81-8.189 11.5-15.22 11.58-9.85 0-15.131-7.26-15.131-21 0-6.63 0-13.89.207-22.05a2.565 2.565 0 0 0-.89-2.43l-4.895-4.95v-.21h21.925v27.39c0 9 1.306 15 6.883 15a8.89 8.89 0 0 0 7.18-4.53v-30a2.838 2.838 0 0 0-.89-2.52l-4.925-5.13v-.21h22.103v42.18zm75.241 7.65c0-1.62.297-22.26.297-27 0-9-1.395-15.21-7.655-15.21a9.911 9.911 0 0 0-7.862 4.32c.628 2.805.926 5.675.89 8.55 0 6.24-.208 14.19-.297 22.14a2.565 2.565 0 0 0 .89 2.43l4.985 4.62v.21h-22.311c0-1.5.296-22.26.296-27 0-9.15-1.394-15.21-7.565-15.21a8.886 8.886 0 0 0-7.358 4.44v30.6c0 1.32 0 1.62.89 2.43l4.866 4.62v.21H295.77v-42a2.838 2.838 0 0 0-.89-2.52l-4.896-5.25v-.21h22.104v10.68c2.05-7.05 8.58-11.788 15.843-11.49 7.358 0 12.342 4.14 14.449 12 2.34-7.249 9.074-12.113 16.614-12 9.85 0 15.428 7.35 15.428 21.24 0 6.63-.207 14.19-.296 22.14a2.412 2.412 0 0 0 1.008 2.43l4.866 4.62v.21h-22.222zm-277.672-6.54l6.557 6.33v.21H53.404v-.21l6.676-6.33a2.565 2.565 0 0 0 .89-2.43v-40.47c0-1.71 0-4.02.296-6L38.836 75h-.266l-23.231-51.75c-.505-1.29-.653-1.38-.98-2.25v33.93a11.905 11.905 0 0 0 1.099 6.63l9.345 12.42v.21H0v-.21l9.346-12.45a11.936 11.936 0 0 0 1.097-6.6v-37.5a7.316 7.316 0 0 0-1.542-5.13L2.285 3.63v-.21H26.02l19.878 43.83L63.403 3.42h23.23v.21l-6.527 7.35a2.594 2.594 0 0 0-1.009 2.52v51.72a2.412 2.412 0 0 0 1.01 2.43zm26.02-20.28v.21c0 12.84 6.557 18 14.241 18 6.416.221 12.355-3.417 15.131-9.27h.208c-2.67 12.69-10.444 18.69-22.074 18.69-12.55 0-24.388-7.65-24.388-25.38 0-18.63 11.868-26.07 24.981-26.07 10.563 0 21.807 5.04 21.807 21.24v2.58h-29.906zm0-2.58h14.627v-2.61c0-13.2-2.7-16.71-6.765-16.71-4.925 0-7.921 5.25-7.921 19.32h.059z'
-      fillRule='nonzero'
-    />
-  </Svg>
-)
-
-const Medium = styled(MediumBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #1c1b1a;
-      }
-    `
-  })};
-`
-
-const SoundCloudBase = props => (
-  <Svg height={118.5} viewBox='28.432 71.063 194.75 111.365' {...props}>
-    <path
-      className='logo'
-      d='M39.375 114.578c-.413 0-.752.332-.804.79l-1.903 17.342 1.903 16.959c.052.45.391.782.804.782.406 0 .738-.324.804-.782l2.154-16.959-2.154-17.35c-.067-.457-.399-.782-.804-.782zm-7.141 6.647c-.406 0-.723.31-.775.752l-1.416 10.733 1.416 10.542c.052.441.369.752.775.752.391 0 .708-.311.774-.752l1.667-10.542-1.667-10.733c-.066-.443-.383-.752-.774-.752zm14.724-10.047c-.509 0-.915.398-.967.951l-1.8 20.581 1.8 19.821c.052.554.458.951.967.951.501 0 .907-.397.974-.943l2.043-19.829-2.043-20.581c-.067-.553-.473-.951-.974-.951zm7.649-.723c-.598 0-1.084.479-1.129 1.114l-1.704 21.142 1.704 20.456c.044.634.531 1.112 1.129 1.112s1.084-.479 1.136-1.112l1.934-20.456-1.934-21.149c-.051-.628-.538-1.107-1.136-1.107zm9.007 42.88zm0-40.247c-.044-.723-.604-1.269-1.291-1.269-.701 0-1.254.546-1.298 1.284l-1.608 19.607 1.608 20.625c.044.729.598 1.284 1.298 1.284.687 0 1.247-.555 1.291-1.284l1.83-20.625-1.83-19.622zm6.485-13.728c-.782 0-1.424.634-1.461 1.453l-1.512 31.912 1.512 20.617c.037.805.679 1.446 1.461 1.446.767 0 1.409-.644 1.46-1.454v.008l1.704-20.617-1.704-31.919c-.052-.811-.693-1.446-1.46-1.446zm7.715-7.355c-.878 0-1.593.716-1.63 1.616l-1.409 39.215 1.409 20.492c.037.893.752 1.607 1.63 1.607.863 0 1.586-.715 1.623-1.615v.008l1.601-20.492-1.601-39.215c-.037-.9-.759-1.616-1.623-1.616zm8.019-3.54c-.974 0-1.763.782-1.792 1.785l-1.321 42.482 1.321 20.277c.029.988.819 1.771 1.792 1.771.959 0 1.748-.782 1.785-1.777v.016l1.482-20.287-1.482-42.482c-.037-1.003-.826-1.785-1.785-1.785zm9.914.354c-.037-1.092-.893-1.948-1.955-1.948-1.07 0-1.933.856-1.955 1.948l-1.224 43.914 1.224 20.16a1.959 1.959 0 0 0 1.955 1.934c1.062 0 1.918-.855 1.955-1.947l1.372-20.146-1.372-43.915zm0 64.059v.015zm6.064-65.048c-1.158 0-2.095.93-2.117 2.11l-1.122 42.792 1.122 19.969c.022 1.173.959 2.104 2.117 2.104s2.088-.931 2.125-2.117v.021l1.254-19.976-1.254-42.792c-.038-1.181-.967-2.111-2.125-2.111zm8.085 1.402c-1.254 0-2.257 1.003-2.287 2.279l-1.018 41.227 1.024 19.837c.022 1.262 1.025 2.265 2.279 2.265s2.265-1.003 2.287-2.279v.016l1.15-19.836-1.15-41.235a2.288 2.288 0 0 0-2.285-2.274zm9.501-7.584a2.451 2.451 0 0 0-1.35-.413c-.487 0-.944.147-1.328.398a2.477 2.477 0 0 0-1.122 2.043l-.015.465-.915 48.589c0 .029.929 19.719.929 19.719v.08c.03.555.236 1.062.576 1.461a2.433 2.433 0 0 0 1.874.895c.664 0 1.269-.281 1.711-.724s.723-1.048.738-1.72l.103-1.955.93-17.748-1.033-49.062a2.512 2.512 0 0 0-1.098-2.028zm1.099 70.802v-.009.009-.014.005zm6.992-75.449a2.557 2.557 0 0 0-2.979.222 2.627 2.627 0 0 0-.966 2.006l-.007.273-1.068 53.245.546 9.841.531 9.582c.015 1.416 1.188 2.59 2.61 2.59 1.425 0 2.598-1.174 2.619-2.611v.021l1.158-19.423-1.158-53.518a2.66 2.66 0 0 0-1.286-2.228zm71.94 30.415c-3.254 0-6.359.656-9.191 1.844-1.889-21.407-19.844-38.197-41.737-38.197a42.468 42.468 0 0 0-15.203 2.833c-1.793.693-2.265 1.409-2.286 2.796v75.397c.021 1.453 1.148 2.582 2.566 2.722.059.008 65.852 0 65.852 0 13.123 0 23.754-10.512 23.754-23.635-.001-13.123-10.632-23.76-23.755-23.76zm-150.42 61.689c-2.781-.655-3.533-1.01-3.533-2.108 0-.774.627-1.579 2.508-1.579 1.608 0 2.862.648 3.991 1.809l2.538-2.457c-1.66-1.732-3.667-2.767-6.403-2.767-3.459 0-6.27 1.962-6.27 5.149 0 3.459 2.257 4.492 5.496 5.236 3.312.752 3.91 1.262 3.91 2.391 0 1.328-.981 1.91-3.062 1.91-1.682 0-3.26-.582-4.492-2.014l-2.53 2.258c1.328 1.955 3.888 3.187 6.823 3.187 4.788 0 6.905-2.25 6.905-5.593-.001-3.82-3.092-4.763-5.881-5.422zm16.163-7.102c-4.788 0-7.62 3.718-7.62 9.06 0 5.349 2.833 9.059 7.62 9.059 4.795 0 7.627-3.71 7.627-9.059 0-5.342-2.832-9.06-7.627-9.06zm0 14.636c-2.833 0-4.013-2.441-4.013-5.576s1.18-5.569 4.013-5.569c2.84 0 4.02 2.435 4.02 5.569 0 3.132-1.18 5.576-4.02 5.576zm21.806-4.493c0 2.803-1.38 4.543-3.607 4.543-2.235 0-3.593-1.785-3.593-4.596v-9.812h-3.511v9.863c0 5.119 2.862 7.975 7.104 7.975 4.463 0 7.119-2.936 7.119-8.025v-9.812h-3.511l-.001 9.864zm18.669-2.434c0 .974.044 3.187.044 3.836-.221-.45-.796-1.327-1.225-1.983l-6.227-9.28h-3.363v17.563h3.467v-7.73c0-.98-.052-3.188-.052-3.836.229.449.804 1.328 1.225 1.984l6.455 9.582h3.135v-17.563h-3.46v7.427zm13.471-7.429h-5.488v17.564h5.237c4.516 0 8.986-2.611 8.986-8.787 0-6.423-3.719-8.777-8.735-8.777zm-.251 14.149h-1.732V165.7H119c3.71 0 5.245 1.776 5.245 5.362-.002 3.215-1.691 5.37-5.379 5.37zm19.851-10.94c1.578 0 2.582.708 3.217 1.902l3.252-1.496c-1.105-2.287-3.104-3.896-6.418-3.896-4.594 0-7.803 3.719-7.803 9.059 0 5.549 3.082 9.06 7.678 9.06 3.188 0 5.289-1.476 6.5-3.97l-2.967-1.756c-.929 1.558-1.852 2.244-3.438 2.244-2.633 0-4.168-2.42-4.168-5.578 0-3.26 1.513-5.569 4.147-5.569zm13.027-3.209h-3.512v17.564h10.562v-3.461h-7.053v-14.104h.003zm16.295-.28c-4.788 0-7.629 3.718-7.629 9.06 0 5.349 2.841 9.059 7.629 9.059 4.795 0 7.635-3.71 7.635-9.059 0-5.342-2.84-9.06-7.635-9.06zm0 14.636c-2.833 0-4.014-2.441-4.014-5.576s1.181-5.569 4.014-5.569c2.84 0 4.014 2.435 4.014 5.569 0 3.132-1.174 5.576-4.014 5.576zm21.807-4.493c0 2.803-1.373 4.543-3.607 4.543-2.235 0-3.594-1.785-3.594-4.596v-9.812h-3.51v9.863c0 5.119 2.861 7.975 7.104 7.975 4.463 0 7.125-2.936 7.125-8.025v-9.812h-3.518v9.864zm13.402-9.863h-5.496v17.564h5.236c4.529 0 8.986-2.611 8.986-8.787.001-6.423-3.703-8.777-8.726-8.777zm-.26 14.149h-1.725V165.7h1.857c3.711 0 5.236 1.776 5.236 5.362.003 3.215-1.67 5.37-5.368 5.37z'
-    />
-  </Svg>
-)
-
-const SoundCloud = styled(SoundCloudBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #ff8800;
-      }
-    `
-  })};
-`
-
-const GizmodoBase = props => (
-  <Svg
-    viewBox='0 0 912.30664 130.90668'
-    height={29.7}
-    xmlSpace='preserve'
-    version='1.1'
-    {...props}
-  >
-    <g transform='matrix(1.3333333,0,0,-1.3333333,0,130.90667) scale(0.1)'>
-      <path
-        className='logo'
-        d='M1121.32 117.23v747.993c0 38.843 8.86 68.007 26.58 87.422 17.7 19.421 40.63 29.152 68.7 29.152 28.96 0 52.39-9.629 70.34-28.836 17.92-19.211 26.91-48.441 26.91-87.738V117.23c0-39.277-8.99-68.64-26.91-88.066C1268.99 9.72 1245.56 0 1216.6 0c-27.66 0-50.45 9.832-68.39 29.48-17.93 19.649-26.89 48.915-26.89 87.75m5230.36 665.36c-161.01 0-292-131.004-292-292.008 0-160.988 130.99-291.984 292-291.984 160.99 0 291.99 130.996 291.99 291.984 0 161.004-131 292.008-291.99 292.008m0-782.59c-270.51 0-490.6 220.086-490.6 490.582 0 270.531 220.09 490.606 490.6 490.606 270.53 0 490.58-220.075 490.58-490.606C6842.26 220.086 6622.21 0 6351.68 0M4149.33 782.59c-160.99 0-291.98-131.004-291.98-292.008 0-160.988 130.99-291.984 291.98-291.984 161.02 0 291.98 130.996 291.98 291.984 0 161.004-130.96 292.008-291.98 292.008m0-782.59c-270.51 0-490.58 220.086-490.58 490.582 0 270.531 220.07 490.606 490.58 490.606 270.52 0 490.6-220.075 490.6-490.606C4639.93 220.086 4419.85 0 4149.33 0m1098.74 201c-23.74-1.547-49.28-1.816-76.75-1.816h-128.8v583.425h128.8c27.47 0 53.01-.304 76.75-1.812 85.15-10.18 159.13-57.055 205.61-124.457 32.38-46.93 51.45-103.664 51.58-164.856 0-.16.02-.316.02-.484v-.215c0-.156-.02-.316-.02-.472-.13-61.192-19.2-117.93-51.58-164.872-46.48-67.398-120.46-114.25-205.61-124.441m455.83 289.898v.133c-.06 108.27-35.36 208.399-94.94 289.641-86.67 118.152-224.8 196.141-381.09 200.738-4.85.149-9.69.387-14.59.387l-246.35-.317c-42.06 0-72.52-9.582-91.34-28.734-18.85-19.156-28.27-50.109-28.27-92.891V121.941c0-42.78 9.42-73.726 28.27-92.902C4894.41 9.875 4924.87.305 4966.93.305L5213.28 0c4.9 0 9.74.238 14.59.371 156.29 4.613 294.42 82.598 381.09 200.754 59.58 81.242 94.88 181.371 94.94 289.637v.136M3380.91 977.422c-35.85 11.137-74.58-2.953-95.46-34.766l-295.8-450.586-295.79 450.586c-20.86 31.813-59.61 45.903-95.46 34.766-35.84-11.149-60.31-44.934-60.31-83.231V93.363c0-48.054 38.11-87.004 85.09-87.004 47.01 0 85.11 38.95 85.11 87.004v515.024L2918.31 288.5c.14-.223.28-.426.42-.652l.26-.383c.25-.395.55-.746.81-1.141 1.31-1.914 2.66-3.82 4.15-5.644.65-.813 1.33-1.535 2-2.301 1.18-1.344 2.32-2.695 3.57-3.969 1.12-1.148 2.32-2.222 3.51-3.324.86-.781 1.72-1.613 2.62-2.379 1.68-1.41 3.38-2.672 5.14-3.914.5-.348.94-.766 1.45-1.105.06-.055.13-.079.22-.133 1.87-1.274 3.81-2.426 5.79-3.551.54-.316 1.08-.688 1.63-.969 1.47-.812 2.99-1.48 4.53-2.199 8.61-4.023 17.63-6.539 26.71-7.477.08 0 .13-.007.18-.031 1.81-.156 3.61-.191 5.39-.269.99-.036 1.99-.125 2.96-.125 1.01 0 1.98.089 2.97.125 1.81.078 3.6.113 5.4.269.05.024.12.031.17.031 9.1.938 18.11 3.454 26.73 7.477 1.52.719 3.06 1.387 4.52 2.199.57.281 1.08.653 1.65.969 1.97 1.125 3.9 2.277 5.77 3.551.09.054.17.078.24.133.5.339.94.757 1.43 1.105 1.76 1.242 3.49 2.504 5.15 3.914.91.766 1.75 1.598 2.65 2.41 1.16 1.071 2.35 2.145 3.47 3.27 1.27 1.297 2.41 2.66 3.58 4.004.68.754 1.37 1.496 2 2.289 1.48 1.824 2.83 3.73 4.15 5.644.26.395.56.746.82 1.141l.24.383c.15.226.28.429.44.652l209.99 319.887V93.363c0-48.054 38.08-87.004 85.14-87.004 46.96 0 85.1 38.95 85.1 87.004v800.828c0 38.297-24.5 72.082-60.35 83.231M2248.45 29.852h-636.17c-33.46 0-70.64 20.933-84.41 53.511-13.78 32.547-16.6 76.864 13.03 112.18l520.39 603.797h-429.92c-47 0-85.12 40.707-85.12 90.93 0 50.222 38.12 80.894 85.12 80.894h579.74c33.45 0 68.1-5.32 98.37-38.484 27.66-30.305 26.44-95.711-4.54-134.715l-502.81-586.277h446.32c47.01 0 81.23-40.708 81.23-90.907 0-50.21-34.22-90.93-81.23-90.93M487.047 0C345.879 0 217.852 55.973 126.551 157.609 46.113 247.152 0 365.137 0 481.316 0 845.91 303.078 981.785 494.133 981.785c187.84 0 325.73-92.113 331.504-96.015 42.304-28.711 53.625-86.723 25.285-129.571-28.274-42.703-85.274-54.246-127.539-25.859-2.465 1.578-100.281 64.672-229.25 64.672-119.754 0-284.895-79.035-284.895-313.696 0-133.496 99.301-290.949 281.367-290.949 102.438 0 174.458 27.293 216.325 53.934v138.847H565.066c-50.921 0-92.207 41.801-92.207 93.387 0 51.574 41.286 93.387 92.207 93.387h234.079c50.918 0 92.203-41.813 92.203-93.387V198.77c0-25.266-10.102-49.422-27.985-67.02C801.543 70.977 665.895 0 487.047 0'
-      />
-    </g>
-  </Svg>
-)
-
-const Gizmodo = styled(GizmodoBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #100f0d;
-      }
-    `
-  })};
-`
-
-const TwitterBase = props => (
-  <Svg
-    xmlns='http://www.w3.org/2000/svg'
-    height={41.5}
-    viewBox='0 0 468.624 93.743'
-    {...props}
-  >
-    <path
-      className='logo'
-      d='M25.254 92.757c-9.128-2.46-17.504-9.394-21.79-18.039C.182 68.093.185 68.14.178 37.69c-.39-26.34-.24-27.19 1.515-29.73C5.422 1.74 13.646.412 19.038 5.159c3.785 3.332 4.086 4.463 4.086 15.367v9.455H39.03c9.728 0 16.83.199 18.29.511 8.3 1.78 11.45 11.706 5.687 17.919-.679.732-2.078 1.786-3.108 2.342-1.809.976-2.48 1.016-19.324 1.154l-17.45.144v4.216c0 6.133.636 8.14 3.483 10.987 3.324 3.324 4.212 3.48 19.796 3.493 12.09.009 13.178.079 15.134.964 7.423 3.36 9.272 12.306 3.799 18.377-3.2 3.549-3.41 3.584-21.131 3.564-13.832-.016-16.08-.122-18.951-.895zm70.824.2c-7.207-1.745-14.645-7.22-19.19-14.127-4.19-6.364-5.007-10.841-5.028-27.567-.019-14.423.27-16.24 3.046-19.097 2.568-2.644 4.204-3.33 7.983-3.344 2.885-.011 3.613.172 5.609 1.405 1.42.878 2.833 2.29 3.71 3.71 1.758 3.046 1.436 4.323 1.653 16.852.23 14.26.257 14.597 1.332 16.327 3.634 5.844 11.148 6.706 16.132 1.849 2.877-2.804 3.148-4.451 3.148-19.179 0-12.11.054-12.924.99-14.788 1.873-3.733 4.59-5.71 8.605-6.26 4.296-.589 8.93 1.944 11.15 6.095.99 1.85 1.049 2.61 1.272 16.264-.08 11.36.455 14.684 1.526 16.275 3.429 5.176 10.634 6.29 15.134 2.338 3.225-2.831 3.223-2.822 3.489-18.649.228-13.585.29-14.383 1.272-16.228 4.12-7.745 14.553-8.233 19.088-.893 2.934 4.067 1.65 12.674 1.65 18.224.212 7.303-.216 16.668-1.493 19.554-3.474 9.807-11.213 17.595-20.522 20.656-4.224 1.389-13.53 1.722-17.722.635-4.043-1.048-7.541-2.654-10.87-4.989l-2.795-1.96-2.794 1.96c-3.38 2.37-6.8 3.975-10.538 4.946-3.485.904-12.09.9-15.837-.008zm97.293.07c-2.134-.93-4.8-3.44-5.964-5.616-1.018-1.903-1.027-2.113-1.027-26.044 0-23.028.042-24.21.92-25.979 3.648-7.357 13.067-9.012 18.778-3.3 3.553 3.551 3.543 3.464 3.39 30.036-.13 22.935-.152 23.403-1.146 25.243-1.309 2.42-3.685 4.68-5.882 5.597-2.125.887-7.098.92-9.07.062zm49.16-.265c-10.414-2.73-19.48-10.956-23.308-21.15-2.066-5.5-2.23-8.303-2.072-35.541.145-25.123.163-25.56 1.158-27.405 4.962-9.195 18.543-8.03 21.489 1.842.496 1.666.689 4.692.689 10.848v8.539l11.126.17c10.287.156 11.266.244 13 1.17 8.141 4.354 8.178 15.155.066 19.49-1.819.972-2.64 1.047-13.105 1.193l-11.164.155.156 5.467c.181 6.38.908 8.292 4.073 10.718 2.772 2.125 4.517 2.472 12.478 2.48 6.387.006 7.242.108 9.395 1.11 8.997 4.19 8.988 16.876-.016 20.966-2.783 1.264-19.08 1.23-23.966-.052zm59.639.023c-11.797-3.227-21.482-13.137-24.537-25.106-.751-2.943-.834-5.913-.837-29.968.178-24.97-1.061-23.994 1.277-29.133 2.693-5.106 8.25-7.356 14.133-5.723 2.453.682 6.58 4.808 7.261 7.261.284 1.021.51 5.814.51 10.847l.003 9.018h9.54c10.926 0 13.827.405 16.46 2.296 5.7 4.095 5.83 13.272.245 17.318-2.787 2.018-5.422 2.403-16.43 2.403h-9.815v4.243c0 5.589.642 7.938 2.853 10.45 3.016 3.426 4.134 3.753 13.816 4.034 8.427.245 8.483.253 10.775 1.668 2.624 1.62 5.052 5.202 5.523 8.148.682 4.264-1.552 9.066-5.406 11.618l-1.958 1.298-10.108.101c-8.628.087-10.575-.026-13.305-.773zm58.475.212c-6.848-1.85-11.895-4.889-16.834-10.14-6-6.377-8.706-13.407-8.688-22.574.01-5.6.97-9.977 3.2-14.609 2.356-4.892 4.94-8.082 8.747-10.799 11.084-7.909 30.785-8.403 43.194-1.083 4.443 2.62 9.062 7.579 11.312 12.144 4.552 9.23 4.094 17.59-1.112 20.309-1.801.94-2.641.983-21.666 1.11l-19.792.132v1.229c0 1.916 3.05 4.512 6.407 5.454 2.289.642 5.05.783 15.406.785 7.007 0 13.371.212 14.313.474 8.064 2.24 9.68 12.92 2.584 17.078-2.298 1.554-3.685 1.163-18.374 1.153-12.178-.023-16.953-.192-18.697-.663zm20.595-41.584c-.882-2.956-3.41-4.706-8.034-5.56-4.918-.91-11.795 1.125-13.121 3.881-.584 1.214-1.142 2.596-1.142 2.596s4.092.137 11.352.137h11.26zm35.351 41.376c-2.414-1.312-4.985-4.358-5.598-6.632-.358-1.33-.472-6.53-.355-16.071.172-13.897.194-14.163 1.457-17.954 3.326-9.986 12.22-18.268 22.566-21.014 3.803-1.009 21.36-1.493 25.189-.694 5.318 1.11 8.774 5.214 8.774 10.42 0 4.44-2.018 7.877-5.798 9.874-1.78.94-2.801 1.04-12.94 1.268-10.514.236-11.089.296-12.776 1.345-2.227 1.385-3.85 3.52-4.509 5.928-.308 1.13-.512 6.717-.514 14.08-.003 12.201-.003 12.205-1.22 14.648-.823 1.65-1.926 2.934-3.397 3.952-1.913 1.324-2.607 1.524-5.692 1.634-2.76.099-3.872-.07-5.187-.784zM192.796 21.741c-7.437-3.706-8.789-13.494-2.613-18.916C192.59.713 194.525 0 197.857 0c4.66 0 8.28 2.177 10.373 6.24 5.005 9.711-5.626 20.388-15.434 15.5z'
-    />
-  </Svg>
-)
-
-const Twitter = styled(TwitterBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #1da1f2;
-      }
-    `
-  })};
-`
-
-const YouTubeBase = props => (
-  <Svg
-    xmlns='http://www.w3.org/2000/svg'
-    height={46.1}
-    viewBox='0 0 380.9 85'
-    fill='black'
-    {...props}
-  >
-    <path
-      className='logo'
-      d='M118.9 13.3c-1.4-5.2-5.5-9.3-10.7-10.7C98.7 0 60.7 0 60.7 0s-38 0-47.5 2.5C8.1 3.9 3.9 8.1 2.5 13.3 0 22.8 0 42.5 0 42.5s0 19.8 2.5 29.2C3.9 76.9 8 81 13.2 82.4 22.8 85 60.7 85 60.7 85s38 0 47.5-2.5c5.2-1.4 9.3-5.5 10.7-10.7 2.5-9.5 2.5-29.2 2.5-29.2s.1-19.8-2.5-29.3z'
-    />
-    <path fill='white' d='M48.6 60.7l31.6-18.2-31.6-18.2z' />
-    <g>
-      <path d='M176.3 77.4c-2.4-1.6-4.1-4.1-5.1-7.6-1-3.4-1.5-8-1.5-13.6v-7.7c0-5.7.6-10.3 1.7-13.8 1.2-3.5 3-6 5.4-7.6 2.5-1.6 5.7-2.4 9.7-2.4 3.9 0 7.1.8 9.5 2.4 2.4 1.6 4.1 4.2 5.2 7.6 1.1 3.4 1.7 8 1.7 13.8v7.7c0 5.7-.5 10.2-1.6 13.7-1.1 3.4-2.8 6-5.2 7.6-2.4 1.6-5.7 2.4-9.8 2.4-4.2-.1-7.6-.9-10-2.5zm13.5-8.4c.7-1.7 1-4.6 1-8.5V43.9c0-3.8-.3-6.6-1-8.4-.7-1.8-1.8-2.6-3.5-2.6-1.6 0-2.8.9-3.4 2.6-.7 1.8-1 4.6-1 8.4v16.6c0 3.9.3 6.8 1 8.5.6 1.7 1.8 2.6 3.5 2.6 1.6 0 2.7-.8 3.4-2.6zM360.9 56.3V59c0 3.4.1 6 .3 7.7.2 1.7.6 3 1.3 3.7.6.8 1.6 1.2 3 1.2 1.8 0 3-.7 3.7-2.1.7-1.4 1-3.7 1.1-7l10.3.6c.1.5.1 1.1.1 1.9 0 4.9-1.3 8.6-4 11s-6.5 3.6-11.4 3.6c-5.9 0-10-1.9-12.4-5.6-2.4-3.7-3.6-9.4-3.6-17.2v-9.3c0-8 1.2-13.8 3.7-17.5 2.5-3.7 6.7-5.5 12.6-5.5 4.1 0 7.3.8 9.5 2.3 2.2 1.5 3.7 3.9 4.6 7 .9 3.2 1.3 7.6 1.3 13.2v9.1h-20.1zm1.5-22.4c-.6.8-1 2-1.2 3.7-.2 1.7-.3 4.3-.3 7.8v3.8h8.8v-3.8c0-3.4-.1-6-.3-7.8-.2-1.8-.7-3-1.3-3.7-.6-.7-1.6-1.1-2.8-1.1-1.4-.1-2.3.3-2.9 1.1zM147.1 55.3L133.5 6h11.9l4.8 22.3c1.2 5.5 2.1 10.2 2.7 14.1h.3c.4-2.8 1.3-7.4 2.7-14l5-22.4h11.9L159 55.3v23.6h-11.8V55.3zM241.6 25.7V79h-9.4l-1-6.5h-.3c-2.5 4.9-6.4 7.4-11.5 7.4-3.5 0-6.1-1.2-7.8-3.5-1.7-2.3-2.5-5.9-2.5-10.9V25.7h12v39.1c0 2.4.3 4.1.8 5.1s1.4 1.5 2.6 1.5c1 0 2-.3 3-1 1-.6 1.7-1.4 2.1-2.4V25.7h12zM303.1 25.7V79h-9.4l-1-6.5h-.3c-2.5 4.9-6.4 7.4-11.5 7.4-3.5 0-6.1-1.2-7.8-3.5-1.7-2.3-2.5-5.9-2.5-10.9V25.7h12v39.1c0 2.4.3 4.1.8 5.1s1.4 1.5 2.6 1.5c1 0 2-.3 3-1 1-.6 1.7-1.4 2.1-2.4V25.7h12z' />
-      <path d='M274.2 15.7h-11.9v63.2h-11.7V15.7h-11.9V6h35.5v9.7zM342.8 34.2c-.7-3.4-1.9-5.8-3.5-7.3s-3.9-2.3-6.7-2.3c-2.2 0-4.3.6-6.2 1.9-1.9 1.2-3.4 2.9-4.4 4.9h-.1V3.3h-11.6v75.6h9.9l1.2-5h.3c.9 1.8 2.3 3.2 4.2 4.3 1.9 1 3.9 1.6 6.2 1.6 4.1 0 7-1.9 8.9-5.6 1.9-3.7 2.9-9.6 2.9-17.5v-8.4c-.1-6.1-.4-10.8-1.1-14.1zm-11 21.7c0 3.9-.2 6.9-.5 9.1-.3 2.2-.9 3.8-1.6 4.7-.8.9-1.8 1.4-3 1.4-1 0-1.9-.2-2.7-.7-.8-.5-1.5-1.2-2-2.1V38.1c.4-1.4 1.1-2.6 2.1-3.6 1-.9 2.1-1.4 3.2-1.4 1.2 0 2.2.5 2.8 1.4.7 1 1.1 2.6 1.4 4.8.3 2.3.4 5.5.4 9.6v7z' />
-    </g>
-  </Svg>
-)
-
-const YouTube = styled(YouTubeBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #ff0000;
-      }
-    `
-  })};
-`
-
-const InstagramBase = props => (
-  <Svg
-    xmlns='http://www.w3.org/2000/svg'
-    height={74}
-    viewBox='0 0 840 300'
-    {...props}
-  >
-    <path
-      className='logo'
-      fill='#262626'
-      d='M64.76 49.108c-16.21 6.785-34.03 25.955-39.659 50.017-7.128 30.487 22.534 43.38 24.967 39.152 2.864-4.967-5.315-6.646-6.998-22.465-2.176-20.43 7.322-43.26 19.276-53.276 2.222-1.861 2.116.73 2.116 5.52 0 8.568-.473 85.48-.473 101.531 0 21.72-.897 28.579-2.508 35.356-1.633 6.868-4.257 11.51-2.269 13.297 2.222 2 11.709-2.757 17.2-10.422 6.587-9.193 8.892-20.234 9.306-32.223.5-14.454.478-37.388.5-50.47.02-11.997.202-47.129-.21-68.249-.102-5.18-14.469-10.615-21.249-7.774m575.628 101.653c-.523 11.292-3.022 20.118-6.124 26.343-6.006 12.052-18.468 15.794-23.759-1.53-2.883-9.445-3.018-25.217-.945-38.396 2.11-13.426 8.003-23.565 17.76-22.65 9.623.903 14.128 13.304 13.068 36.233zM478.15 220.811c-.13 18.762-3.083 35.212-9.415 39.99-8.981 6.774-21.052 1.692-18.553-11.998 2.212-12.115 12.671-24.487 27.994-39.605 0 0 .032 3.448-.025 11.613zm-2.453-70.165c-.549 10.284-3.215 20.616-6.124 26.453-6.004 12.052-18.556 15.82-23.758-1.53-3.556-11.855-2.705-27.198-.945-36.865 2.284-12.543 7.819-24.18 17.76-24.18 9.664 0 14.43 10.603 13.067 36.123zm-93.983-.158c-.59 10.893-2.714 19.998-6.124 26.611-6.17 11.97-18.375 15.765-23.758-1.53-3.88-12.472-2.56-29.478-.945-38.664 2.395-13.632 8.394-23.297 17.759-22.38 9.62.939 14.296 13.302 13.068 35.964zm430.704 12.723c-2.352 0-3.425 2.424-4.313 6.504-3.079 14.187-6.316 17.39-10.49 17.39-4.663 0-8.853-7.025-9.93-21.087-.846-11.057-.71-31.414.372-51.663.222-4.16-.926-8.277-12.085-12.331-4.802-1.744-11.782-4.311-15.257 4.078-9.82 23.7-13.662 42.517-14.566 50.16-.047.397-.532.477-.617-.447-.576-6.122-1.863-17.247-2.023-40.621-.031-4.561-.998-8.443-6.031-11.621-3.266-2.062-13.184-5.71-16.756-1.37-3.094 3.548-6.677 13.098-10.4 24.419a2056.41 2056.41 0 0 1-5.134 15.425s.04-24.826.076-34.243c.016-3.552-2.421-4.736-3.155-4.95-3.308-.96-9.824-2.564-12.59-2.564-3.413 0-4.249 1.907-4.249 4.687 0 .364-.538 32.686-.538 55.286 0 .982 0 2.053.01 3.191-1.888 10.388-8.01 24.489-14.667 24.489-6.665 0-9.81-5.896-9.81-32.84 0-15.718.472-22.552.704-33.92.135-6.549.395-11.577.38-12.718-.049-3.5-6.1-5.263-8.916-5.915-2.828-.657-5.287-.912-7.206-.803-2.717.154-4.64 1.936-4.64 4.387 0 1.315.016 3.814.016 3.814-3.5-5.499-9.128-9.326-12.872-10.435-10.086-2.995-20.611-.341-28.55 10.767-6.31 8.827-10.112 18.823-11.61 33.186-1.094 10.5-.737 21.148 1.208 30.153-2.35 10.16-6.713 14.324-11.49 14.324-6.937 0-11.965-11.322-11.381-30.901.385-12.878 2.962-21.916 5.779-34.99 1.2-5.574.225-8.492-2.223-11.29-2.244-2.564-7.027-3.874-13.904-2.263-4.897 1.15-11.9 2.385-18.307 3.334 0 0 .387-1.542.704-4.261 1.667-14.262-13.83-13.105-18.775-8.55-2.951 2.72-4.96 5.927-5.723 11.694-1.209 9.152 6.255 13.468 6.255 13.468-2.449 11.21-8.452 25.854-14.65 36.441-3.318 5.673-5.858 9.876-9.136 14.344-.012-1.664-.021-3.327-.027-4.983-.075-23.568.239-42.116.377-48.802.135-6.548.397-11.443.383-12.584-.037-2.56-1.532-3.526-4.64-4.75-2.75-1.082-6-1.831-9.37-2.092-4.256-.332-6.82 1.925-6.753 4.593.013.504.013 3.598.013 3.598-3.5-5.499-9.128-9.326-12.872-10.435-10.086-2.994-20.61-.34-28.55 10.767-6.307 8.827-10.439 21.212-11.61 33.077-1.09 11.057-.888 20.454.598 28.37-1.604 7.927-6.215 16.216-11.428 16.216-6.666 0-10.458-5.896-10.458-32.84 0-15.718.472-22.552.704-33.92.134-6.548.394-11.577.38-12.717-.05-3.5-6.1-5.263-8.916-5.916-2.946-.681-5.49-.932-7.443-.788-2.577.192-4.388 2.5-4.388 4.22v3.966c-3.5-5.499-9.128-9.326-12.872-10.435-10.086-2.994-20.552-.297-28.55 10.767-5.215 7.215-9.436 15.213-11.61 32.92-.627 5.116-.905 9.907-.869 14.386-2.08 12.718-11.264 27.376-18.777 27.376-4.397 0-8.584-8.528-8.584-26.702 0-24.208 1.498-58.676 1.752-61.998 0 0 9.493-.162 11.33-.183 4.736-.052 9.025.06 15.331-.263 3.164-.161 6.211-11.513 2.946-12.918-1.48-.636-11.937-1.194-16.083-1.282-3.486-.079-13.191-.797-13.191-.797s.87-22.878 1.073-25.295c.172-2.014-2.434-3.051-3.928-3.68-3.634-1.538-6.886-2.273-10.74-3.068-5.324-1.098-7.74-.024-8.211 4.47-.711 6.82-1.079 26.796-1.079 26.796-3.907 0-17.254-.764-21.162-.764-3.632 0-7.551 15.617-2.53 15.81 5.776.223 15.842.416 22.515.618 0 0-.297 35.036-.297 45.852 0 1.151.01 2.26.012 3.335-3.673 19.143-16.61 29.483-16.61 29.483 2.779-12.664-2.896-22.174-13.118-30.225-3.766-2.966-11.2-8.582-19.519-14.737 0 0 4.818-4.748 9.09-14.301 3.028-6.767 3.159-14.51-4.272-16.218-12.278-2.824-22.401 6.193-25.42 15.82-2.34 7.457-1.092 12.99 3.49 18.738.335.42.698.85 1.073 1.281-2.772 5.342-6.58 12.534-9.804 18.112-8.95 15.485-15.711 27.733-20.821 27.733-4.085 0-4.03-12.435-4.03-24.078 0-10.037.741-25.127 1.333-40.75.196-5.168-2.388-8.111-6.719-10.777-2.631-1.62-8.247-4.805-11.5-4.805-4.869 0-18.915.663-32.187 39.054-1.673 4.839-4.959 13.656-4.959 13.656l.283-46.167c0-1.083-.576-2.13-1.896-2.845-2.237-1.214-8.21-3.698-13.522-3.698-2.53 0-3.793 1.177-3.793 3.523l-.464 72.23c0 5.487.143 11.89.686 14.69.54 2.803 1.416 5.084 2.501 6.441 1.083 1.355 2.336 2.388 4.402 2.814 1.922.396 12.45 1.748 12.997-2.275.656-4.822.68-10.037 6.209-29.487 8.606-30.28 19.827-45.055 25.102-50.302.922-.917 1.975-.972 1.925.53-.226 6.638-1.018 23.227-1.55 37.32-1.43 37.716 5.432 44.706 15.236 44.706 7.502 0 18.076-7.453 29.41-26.32a6626.986 6626.986 0 0 0 18.858-31.595c3.436 3.18 7.291 6.603 11.144 10.26 8.953 8.496 11.892 16.57 9.942 24.23-1.492 5.856-7.11 11.89-17.108 6.025-2.914-1.71-4.158-3.033-7.088-4.962-1.574-1.037-3.978-1.346-5.419-.26-3.743 2.822-5.884 6.412-7.106 10.856-1.19 4.325 3.142 6.61 7.632 8.61 3.866 1.722 12.175 3.281 17.474 3.459 20.646.69 37.186-9.97 48.7-37.465 2.06 23.746 10.832 37.183 26.072 37.183 10.189 0 20.404-13.17 24.871-26.125 1.283 5.28 3.18 9.873 5.632 13.757 11.743 18.606 34.524 14.601 45.968-1.198 3.538-4.882 4.076-6.636 4.076-6.636 1.669 14.918 13.683 20.13 20.561 20.13 7.704 0 15.659-3.641 21.234-16.192a38.367 38.367 0 0 0 2.145 3.896c11.743 18.605 34.524 14.6 45.967-1.198a72.898 72.898 0 0 0 1.417-2.017l.336 9.799s-6.528 5.988-10.536 9.662c-17.636 16.18-31.047 28.457-32.033 42.753-1.267 18.229 13.517 25.003 24.702 25.891 11.877.943 22.049-5.619 28.3-14.801 5.5-8.082 9.099-25.478 8.835-42.658-.105-6.88-.279-15.628-.415-25.005 6.2-7.199 13.183-16.297 19.614-26.947 7.008-11.606 14.519-27.192 18.364-39.322 0 0 6.526.056 13.49-.4 2.227-.145 2.867.31 2.455 1.943-.497 1.973-8.798 33.99-1.222 55.32 5.186 14.601 16.875 19.299 23.806 19.299 8.113 0 15.874-6.127 20.034-15.225a32.065 32.065 0 0 0 1.597 2.902c11.743 18.605 34.444 14.577 45.967-1.198 2.6-3.56 4.076-6.637 4.076-6.637 2.473 15.44 14.48 20.21 21.357 20.21 7.164 0 13.963-2.937 19.48-15.99.23 5.748.593 10.447 1.166 11.928.35.906 2.385 2.044 3.866 2.593 6.553 2.43 13.237 1.28 15.71.781 1.713-.347 3.049-1.721 3.231-5.27.48-9.318.186-24.974 3.01-36.609 4.742-19.527 9.165-27.101 11.264-30.851 1.174-2.101 2.5-2.448 2.547-.224.1 4.499.323 17.711 2.16 35.462 1.349 13.056 3.151 20.772 4.536 23.215 3.953 6.984 8.835 7.314 12.81 7.314 2.529 0 7.817-.698 7.344-5.143-.23-2.166.173-15.554 4.848-34.792 3.053-12.563 8.143-23.914 9.978-28.064.677-1.53.992-.324.981-.089-.386 8.652-1.254 36.95 2.27 52.428 4.778 20.966 18.6 23.312 23.416 23.312 10.282 0 18.691-7.821 21.524-28.4.682-4.953-.328-8.778-3.355-8.778'
-    />
-  </Svg>
-)
-
-const Instagram = styled(InstagramBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #30618a;
-      }
-    `
-  })};
-`
-
-const Wikipedia = props => (
-  <Svg
-    xmlns='http://www.w3.org/2000/svg'
-    height={35.2}
-    viewBox='0 0 500 85'
-    {...props}
-  >
-    <path d='M55.274 22.072s-1.468-2.984-1.968-3.976c-3.336-6.612-3.264-6.964-6.636-7.408-.944-.124-1.432-.244-1.432-.712V7.884l.284-.204c6.092-.008 19.648-.008 19.648-.008l.52.172v2.084c0 .476-.34.72-1.032.72l-1.424.204c-3.6.28-3.012 1.752-.62 6.516l22.708 46.46.776.212 20.16-47.844c.7-1.928.592-3.296-.308-4.1-.904-.784-1.54-1.248-3.876-1.344l-1.892-.096a1.11 1.11 0 0 1-.664-.24c-.208-.148-.316-.344-.316-.58V7.844l.288-.204h22.776l.232.204v1.992c0 .54-.308.812-.916.812-2.984.132-5.196.78-6.628 1.936-1.436 1.164-2.556 2.768-3.36 4.852 0 0-18.54 42.436-24.884 56.54-2.428 4.624-4.816 4.2-6.916-.132-4.436-9.136-16.928-36.852-16.928-36.852l-7.592-14.92z' />
-    <path d='M91.658 7.648s-12.812-.04-18.516 0l-.284.204v1.984c0 .244.104.436.312.584.212.156.428.236.66.236l.916.1c2.332.096 3.404.708 3.724 1.092.556.68.832 1.416-.516 4.356l-23.92 47.708-.628-.164S36.274 27.624 31.314 15.54c-.516-1.252-.748-2.18-.748-2.752 0-1.228 1.108-1.892 3.324-1.984l2.584-.108c.684 0 1.032-.264 1.032-.812V7.9l-.232-.204s-20.52-.024-24.76 0l-.232.204v2.084c0 .376.476.608 1.432.712 2.604.14 4.316.556 5.14 1.248.82.7 1.688 2.468 2.84 5.224 6.228 14.96 19.46 41.656 25.904 56.86 1.856 4.152 4.188 4.796 7.02-.128 4.904-9.096 18.2-36.872 18.2-36.872L83.29 17.444c1.22-2.068 2.396-3.912 2.992-4.772 1.092-1.572 1.704-1.868 4.688-2.008.608 0 .916-.272.916-.82V7.86l-.228-.212zM130.614 69.58c0 2.288 1.04 3.64 4.368 3.952l2.184.212c.312 0 .624.204.624.62v2.184l-.104.212s-7.904-.212-11.128-.212c-2.912 0-11.024.212-11.024.212l-.208-.212v-2.184c0-.416.312-.62.728-.62l2.08-.212c3.432-.312 4.472-1.976 4.472-3.952V34.948c0-2.284-1.252-2.916-4.472-3.22l-2.08-.216c-.416 0-.728-.096-.728-.416V28.7l.104-.208s8.216.208 11.44.208c2.808 0 10.816-.208 10.816-.208l.104.208v2.288c0 .312-.208.52-.624.52l-2.184.212c-3.224.308-4.368 1.144-4.368 3.228V69.58zM161.306 69.58c0 2.288 1.352 3.432 4.368 3.952l1.144.212c.312.1.624.204.624.62v2.184l-.104.212s-6.552-.212-9.672-.212c-2.912 0-11.44.212-11.44.212l-.208-.212v-2.184c0-.416.312-.62.728-.62l2.08-.212c3.12-.312 4.472-1.976 4.472-3.952V34.948c0-2.284-1.144-2.916-4.472-3.22l-2.08-.216c-.416 0-.728-.096-.728-.416V28.7l.104-.208s8.528.208 11.752.208c2.916 0 9.464-.208 9.464-.208l.104.208v2.288c0 .312-.208.416-.624.52l-1.144.212c-3.224.62-4.368 1.144-4.368 3.228v15.704c3.432 0 5.408-1.764 9.984-6.656 4.16-4.58 7.384-8.42 7.384-10.4 0-.836-.936-1.768-3.016-2.084-.416-.1-.624-.208-.624-.52v-2.396l.104-.1s8.424.208 11.44.208c2.704 0 7.28-.208 7.28-.208l.312.1v2.396c0 .312-.312.52-.832.52-4.056.312-6.864.52-10.608 4.784l-8.944 10.292c-1.768 1.876-2.392 3.016-2.392 3.852 0 .624.416 1.14.936 1.764l15.704 18.928c1.248 1.452 3.744 2.08 5.824 2.492l.52.1c.312.104.728.312.728.624v2.184l-.312.104s-5.512-.104-8.424-.104c-2.6 0-6.344.104-6.344.104l-.104-.104v-1.248c0-1.356-.104-2.492-1.56-4.156l-11.544-14.356c-1.56-2.084-3.224-2.596-5.512-2.596V69.58zM217.374 69.58c0 2.288 1.04 3.64 4.368 3.952l2.184.212c.312 0 .624.204.624.62v2.184l-.104.212s-7.904-.212-11.128-.212c-2.912 0-11.024.212-11.024.212l-.208-.212v-2.184c0-.416.312-.62.728-.62l2.08-.212c3.432-.312 4.472-1.976 4.472-3.952V34.948c0-2.284-1.248-2.916-4.472-3.22l-2.08-.216c-.416 0-.728-.096-.728-.416V28.7l.104-.208s8.216.208 11.44.208c2.808 0 10.816-.208 10.816-.208l.104.208v2.288c0 .312-.208.52-.624.52l-2.184.212c-3.224.308-4.368 1.144-4.368 3.228V69.58zM239.846 35.152c0-3.224-1.352-3.328-6.24-3.64-.624 0-.728-.308-.728-.52V28.6l.208-.104c1.56 0 7.696.208 10.712.208 2.912 0 7.072-.208 10.92-.208 14.76 0 17.464 7.276 17.464 12.692 0 6.244-3.016 16.016-17.776 16.016-2.6 0-4.68-.212-6.556-1.04v13.928c0 2.288 1.248 2.812 4.58 3.328l2.084.312c.416.104.728.212.728.624v2.184l-.1.208s-8.32-.208-11.544-.208c-2.912 0-10.712.208-10.712.208l-.104-.208v-2.184c0-.416.208-.52.624-.624l2.184-.312c3.224-.416 4.264-1.352 4.264-3.328l-.008-34.94zm8.008 18.1c2.08.836 4.68.936 6.76.936 6.244 0 9.248-3.752 9.248-11.752 0-8.944-4.056-10.92-9.768-10.92-4.996 0-6.244.936-6.244 3.64l.004 18.096zM290.094 76.548c-3.12 0-10.292.212-10.292.212l-.108-.212V74.26c0-.204.108-.516.624-.516l2.396-.212c3.228-.312 4.164-2.392 4.164-4.468v-34.32c0-2.4-1.252-2.812-4.164-3.016l-2.396-.216c-.624-.096-.624-.308-.624-.416V28.7l.108-.208s8.108.208 11.228.208h15.288c3.12 0 10.508-.208 10.508-.208l.212.208c.1 4.472.828 8.536 1.56 12.484l-2.596.416c-1.14-3.852-3.116-9.36-9.772-9.36h-8.32c-1.976 0-3.016.728-3.016 2.492v15.084h7.796c2.916 0 3.64-1.248 4.06-3.328l.62-2.184c0-.312.212-.524.624-.524h1.764c.212 0 .212.108.212.312-.212 2.288-.624 5.204-.624 7.7 0 2.596.416 4.784.624 7.176l-.212.212h-1.764c-.516 0-.732-.212-.732-.524l-.52-2.284c-.416-2.084-1.352-3.228-4.064-3.228h-7.792V69.06c0 1.972 1.136 3.948 3.016 3.948h9.88c5.304 0 8.22-3.328 9.464-8.74l2.596.312-2.084 11.956-.208.208s-6.136-.208-9.256-.208l-18.2.012zM329.206 73.744l2.392-.212c3.124-.312 4.164-1.248 4.164-4.468v-34.32c0-2.712-.736-2.712-4.164-3.016l-2.28-.216c-.416 0-.736-.208-.736-.52v-2.288l.1-.208s9.156.208 11.032.208c1.968 0 11.024-.208 13.728-.208 17.368 0 25.9 10.816 25.9 24.648 0 18.204-12.9 23.604-24.44 23.604-8.32 0-11.956-.208-15.18-.208-3.124 0-11.028.208-11.028.208l-.104-.208v-2.184a.587.587 0 0 1 .616-.612zm23.196 0c11.436 0 17.892-4.26 17.892-18.828 0-10.812-2.396-23.404-19.868-23.404-4.056 0-6.656.624-6.656 3.232v34.32c0 2.904.724 4.68 8.632 4.68zM403.894 69.58c0 2.288 1.04 3.64 4.368 3.952l2.184.212c.312 0 .624.204.624.62v2.184l-.104.212s-7.904-.212-11.124-.212c-2.916 0-11.024.212-11.024.212l-.212-.212v-2.184c0-.416.312-.62.728-.62l2.084-.212c3.428-.312 4.472-1.976 4.472-3.952V34.948c0-2.284-1.252-2.916-4.472-3.22l-2.084-.216c-.416 0-.728-.096-.728-.416V28.7l.104-.208s8.216.208 11.432.208c2.808 0 10.82-.208 10.82-.208l.1.208v2.288c0 .312-.212.52-.624.52l-2.184.212c-3.224.308-4.368 1.144-4.368 3.228l.008 34.632zM435.934 73.532l3.016.212c.416 0 .832.312.832.828v1.976l-.208.212s-11.336-.212-13.832-.212c-2.708 0-9.364.212-9.364.212l-.212-.212v-1.976c0-.516.312-.828.836-.828 4.58-.312 7.172-1.148 10.084-7.804l23.084-54.084c1.14-2.704 1.976-3.748 3.54-3.748 1.452 0 1.976 1.044 2.804 3.124l21.32 56.368c1.46 4.06 3.54 5.616 7.7 6.032l1.144.1c.52 0 1.04.312 1.04.836v1.972l-.312.208s-8.008-.208-12.064-.208c-3.64 0-13.836.208-13.836.208l-.1-.208v-1.972c0-.524.208-.836.736-.836l2.912-.208c3.948-.308 5.412-1.564 4.368-4.472l-6.344-17.58h-21.632c-2.496 0-3.12.416-3.636 1.764l-6.036 15.084c-1.04 2.608.312 4.9 4.16 5.212zm25.48-26.424l-10.088-27.764h-.208l-10.4 26.416c-.416 1.044.308 1.352 2.596 1.352h18.1z' />
-  </Svg>
-)
-
-const Apple = props => (
-  <Svg
-    xmlns='http://www.w3.org/2000/svg'
-    xmlnsXlink='http://www.w3.org/1999/xlink'
-    width={207}
-    viewBox='0 0 2541.3 903.4'
-    {...props}
-  >
-    <defs>
-      <path id='SVGID_1_' d='M0 0h708.7v870.3H0z' />
-    </defs>
-    <clipPath id='SVGID_2_'>
-      <use overflow='visible' xlinkHref='#SVGID_1_' />
-    </clipPath>
-    <g clipPath='url(#SVGID_2_)'>
-      <path
-        id='path188'
-        d='M591.8 462.4c-1-110.1 90-163 94.1-165.6-51.2-74.9-130.8-85.1-159.2-86.3-67.7-6.9-132.2 39.9-166.6 39.9s-87.4-38.9-143.6-37.8c-73.9 1.1-142 42.9-180.1 109.1-76.7 133-19.6 330.3 55.2 438.4 36.6 52.8 80.1 112.3 137.4 110.1 55.2-2.2 76-35.7 142.6-35.7s85.4 35.7 143.6 34.6c59.3-1.2 96.9-54 133.2-107 41.9-61.3 59.2-120.7 60.2-123.8-1.3-.5-115.6-44.3-116.8-175.9'
-      />
-    </g>
-    <path
-      id='path190'
-      d='M482.3 139c30.4-36.8 50.9-88 45.3-139-43.8 1.8-96.8 29.2-128.2 66-28.1 32.5-52.8 84.6-46.1 134.6 48.7 3.8 98.6-24.9 129-61.6'
-    />
-    <g className='st1'>
-      <path d='M975.7 570.7l-50.4 152.7h-64.8L1025.4 238h75.6l165.6 485.4h-67l-51.9-152.7h-172zm159.2-49L1087.4 382c-10.8-31.7-18-60.5-25.2-88.6h-1.4c-7.2 28.8-15.1 58.3-24.5 87.9l-47.5 140.4h146.1z' />
-    </g>
-    <g className='st1'>
-      <path d='M1315.7 488.6c0-44.6-1.4-80.7-2.9-113.8h56.9l2.9 59.8h1.4c25.9-42.5 67-67.7 123.9-67.7 84.3 0 147.6 71.3 147.6 177.2 0 125.3-76.3 187.2-158.4 187.2-46.1 0-86.4-20.2-107.3-54.7h-1.4V866h-62.7V488.6zm62.6 92.9c0 9.4 1.4 18 2.9 25.9 11.5 43.9 49.7 74.2 95.1 74.2 67 0 105.9-54.7 105.9-134.7 0-69.9-36.7-129.6-103.7-129.6-43.2 0-83.5 31-95.8 78.5-2.2 7.9-4.3 17.3-4.3 25.9v59.8zM1711 488.6c0-44.6-1.4-80.7-2.9-113.8h56.9l2.9 59.8h1.4c25.9-42.5 67-67.7 123.9-67.7 84.3 0 147.6 71.3 147.6 177.2 0 125.3-76.3 187.2-158.4 187.2-46.1 0-86.4-20.2-107.3-54.7h-1.4V866H1711V488.6zm62.7 92.9c0 9.4 1.4 18 2.9 25.9 11.5 43.9 49.7 74.2 95.1 74.2 67 0 105.9-54.7 105.9-134.7 0-69.9-36.7-129.6-103.7-129.6-43.2 0-83.5 31-95.8 78.5-2.2 7.9-4.3 17.3-4.3 25.9v59.8zM2106.4 212.1h63.4v511.3h-63.4V212.1zM2297.2 560.6c1.4 85.7 56.2 121 119.5 121 45.4 0 72.7-7.9 96.5-18L2524 709c-22.3 10.1-60.5 21.6-115.9 21.6-107.3 0-171.4-70.6-171.4-175.7s61.9-188 163.5-188c113.8 0 144 100.1 144 164.2 0 13-1.4 23-2.2 29.5h-244.8zm185.8-45.4c.7-40.3-16.6-103-87.9-103-64.1 0-92.2 59.1-97.2 103H2483z' />
-    </g>
-  </Svg>
-)
-
-const AmazonBase = props => (
-  <Svg
-    xmlns='http://www.w3.org/2000/svg'
-    xmlnsXlink='http://www.w3.org/1999/xlink'
-    height={62.5}
-    viewBox='0 0 603 182'
-    fill='black'
-    {...props}
-  >
-    <path
-      className='logo'
-      d='M374.006 142.184c-35 25.797-85.729 39.561-129.406 39.561-61.242 0-116.376-22.651-158.087-60.325-3.278-2.962-.341-7 3.591-4.693 45.015 26.191 100.673 41.947 158.166 41.947 38.775 0 81.43-8.022 120.65-24.67 5.925-2.516 10.88 3.88 5.086 8.18'
-    />
-    <path
-      className='logo'
-      d='M388.557 125.536c-4.457-5.715-29.573-2.7-40.846-1.363-3.434.42-3.959-2.57-.865-4.719 20.003-14.078 52.827-10.015 56.654-5.296 3.828 4.745-.996 37.648-19.793 53.352-2.884 2.411-5.637 1.127-4.352-2.072 4.22-10.539 13.685-34.16 9.202-39.902'
-    />
-    <path d='M348.497 20.066V6.381c0-2.071 1.573-3.46 3.461-3.46h61.269c1.966 0 3.54 1.415 3.54 3.46V18.1c-.027 1.966-1.679 4.535-4.615 8.599l-31.749 45.329c11.798-.289 24.25 1.468 34.947 7.498 2.412 1.363 3.068 3.356 3.251 5.322V99.45c0 1.992-2.202 4.325-4.509 3.12-18.85-9.884-43.887-10.96-64.73.104-2.123 1.154-4.351-1.153-4.351-3.146V85.661c0-2.229.026-6.03 2.254-9.412L384.047 23.5h-32.01c-1.967 0-3.54-1.39-3.54-3.434M124.999 105.454h-18.64c-1.783-.13-3.199-1.468-3.33-3.172V6.617c0-1.914 1.6-3.435 3.592-3.435h17.382c1.809.079 3.25 1.468 3.382 3.199v12.505h.34c4.536-12.086 13.056-17.722 24.54-17.722 11.666 0 18.954 5.636 24.198 17.722 4.509-12.086 14.76-17.722 25.744-17.722 7.813 0 16.36 3.224 21.577 10.46 5.899 8.049 4.693 19.741 4.693 29.992l-.026 60.378c0 1.913-1.6 3.46-3.592 3.46h-18.614c-1.862-.13-3.356-1.625-3.356-3.46V51.29c0-4.037.367-14.104-.524-17.932-1.39-6.423-5.558-8.232-10.959-8.232-4.51 0-9.228 3.015-11.142 7.839-1.914 4.824-1.73 12.898-1.73 18.325v50.704c0 1.913-1.6 3.46-3.592 3.46h-18.614c-1.888-.13-3.356-1.625-3.356-3.46l-.026-50.704c0-10.67 1.757-26.374-11.483-26.374-13.397 0-12.872 15.31-12.872 26.374v50.704c0 1.913-1.6 3.46-3.592 3.46M469.514 1.164c27.66 0 42.629 23.752 42.629 53.954 0 29.18-16.543 52.329-42.629 52.329-27.16 0-41.947-23.753-41.947-53.352 0-29.782 14.97-52.931 41.947-52.931m.158 19.531c-13.738 0-14.603 18.719-14.603 30.386 0 11.692-.184 36.65 14.445 36.65 14.446 0 15.128-20.134 15.128-32.403 0-8.075-.341-17.723-2.78-25.378-2.097-6.66-6.265-9.255-12.19-9.255M548.008 105.454h-18.562c-1.861-.13-3.356-1.625-3.356-3.46l-.026-95.692c.157-1.756 1.704-3.12 3.592-3.12h17.277c1.625.079 2.962 1.18 3.33 2.674v14.63h.34c5.217-13.083 12.532-19.322 25.404-19.322 8.363 0 16.517 3.015 21.76 11.273 4.877 7.655 4.877 20.528 4.877 29.782v60.22c-.21 1.678-1.757 3.015-3.592 3.015h-18.693c-1.704-.13-3.12-1.39-3.303-3.015V50.478c0-10.461 1.206-25.772-11.667-25.772-4.535 0-8.704 3.042-10.775 7.656-2.621 5.846-2.962 11.666-2.962 18.116v51.516c-.026 1.913-1.652 3.46-3.644 3.46' />
-    <use transform='translate(244.367)' xlinkHref='#a' />
-    <path
-      id='a'
-      d='M55.288 59.758v-4.037c-13.475 0-27.71 2.884-27.71 18.771 0 8.049 4.168 13.502 11.325 13.502 5.243 0 9.936-3.225 12.898-8.468 3.67-6.45 3.487-12.506 3.487-19.768m18.798 45.434c-1.232 1.101-3.015 1.18-4.405.446-6.187-5.139-7.288-7.524-10.696-12.427-10.225 10.434-17.46 13.554-30.726 13.554-15.678 0-27.895-9.674-27.895-29.048 0-15.127 8.206-25.43 19.872-30.464 10.12-4.457 24.25-5.244 35.052-6.476v-2.412c0-4.43.341-9.674-2.254-13.501-2.281-3.435-6.633-4.85-10.46-4.85-7.106 0-13.45 3.644-14.997 11.194-.315 1.678-1.547 3.33-3.225 3.408l-18.09-1.94c-1.52-.34-3.198-1.573-2.778-3.906C7.652 6.853 27.446.246 45.169.246c9.07 0 20.92 2.412 28.078 9.28 9.07 8.469 8.206 19.768 8.206 32.064v29.048c0 8.73 3.618 12.558 7.026 17.277 1.206 1.678 1.468 3.697-.053 4.955-3.801 3.172-10.565 9.071-14.288 12.375l-.052-.053'
-    />
-  </Svg>
-)
-
-const Amazon = styled(AmazonBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #f90;
-      }
-    `
-  })};
-`
-
-const VimeoBase = props => (
-  <Svg
-    xmlns='http://www.w3.org/2000/svg'
-    height={59.1}
-    viewBox='0 0 70 20'
-    {...props}
-  >
-    <path
-      className='logo'
-      d='M15.72 9.431c-.069 1.514-1.127 3.588-3.172 6.22-2.114 2.749-3.903 4.124-5.367 4.124-.906 0-1.673-.837-2.3-2.512l-1.255-4.604c-.465-1.674-.964-2.512-1.498-2.512-.116 0-.524.245-1.221.733l-.731-.943C.943 9.263 1.7 8.588 2.445 7.912c1.024-.884 1.792-1.35 2.305-1.397 1.21-.116 1.955.712 2.235 2.483.302 1.912.511 3.101.628 3.566.349 1.586.733 2.378 1.152 2.378.326 0 .815-.515 1.467-1.543.651-1.029 1-1.812 1.047-2.349.093-.888-.256-1.333-1.047-1.333-.373 0-.757.085-1.151.255.764-2.504 2.224-3.721 4.38-3.652 1.598.047 2.351 1.084 2.259 3.111'
-    />
-    <path
-      className='logo'
-      d='M22.281 1.918c-.023.58-.314 1.136-.874 1.669a3.12 3.12 0 0 1-2.234.903c-1.327 0-1.968-.579-1.921-1.737.022-.602.378-1.182 1.064-1.738.687-.555 1.449-.834 2.288-.834.489 0 .896.192 1.223.574.325.382.477.77.454 1.163zm3.038 12.419c-.652 1.232-1.548 2.349-2.689 3.349-1.56 1.349-3.119 2.024-4.679 2.024-.723 0-1.275-.233-1.659-.699-.384-.465-.565-1.069-.541-1.814.022-.767.261-1.954.715-3.56.454-1.605.682-2.466.682-2.582 0-.605-.21-.908-.629-.908-.139 0-.536.245-1.188.733l-.803-.943 2.235-2.025c1.001-.884 1.746-1.35 2.236-1.397.768-.069 1.332.157 1.693.679.36.523.494 1.2.402 2.035-.303 1.415-.629 3.212-.978 5.392-.024.998.337 1.496 1.082 1.496.326 0 .908-.344 1.746-1.033.699-.574 1.269-1.114 1.712-1.62l.663.873'
-    />
-    <path
-      className='logo'
-      d='M47.127 14.336c-.652 1.233-1.548 2.349-2.689 3.349-1.56 1.349-3.12 2.024-4.679 2.024-1.514 0-2.247-.837-2.2-2.513.022-.745.168-1.639.436-2.686.267-1.048.413-1.862.436-2.444.024-.883-.245-1.326-.806-1.326-.607 0-1.331.722-2.172 2.165-.887 1.514-1.367 2.98-1.436 4.4-.05 1.002.05 1.77.293 2.305-1.624.047-2.762-.221-3.411-.803-.582-.512-.848-1.361-.801-2.549.02-.745.136-1.49.343-2.235.205-.745.319-1.408.342-1.991.05-.861-.268-1.292-.944-1.292-.583 0-1.213.664-1.888 1.991a10.141 10.141 0 0 0-1.119 4.155c-.05 1.305.04 2.212.25 2.724-1.598.047-2.733-.29-3.404-1.01-.558-.603-.812-1.52-.765-2.751.02-.603.129-1.445.321-2.524.192-1.08.299-1.921.321-2.525.05-.417-.06-.627-.314-.627-.14 0-.536.236-1.188.707l-.838-.943c.117-.092.849-.768 2.2-2.025.978-.907 1.641-1.373 1.99-1.396.606-.047 1.094.203 1.467.75.372.547.559 1.182.559 1.903 0 .233-.02.454-.07.664a6.433 6.433 0 0 1 1.222-1.398c1.071-.93 2.27-1.455 3.597-1.571 1.141-.093 1.955.174 2.445.803.395.512.581 1.246.558 2.2.163-.139.338-.291.525-.454.534-.628 1.058-1.128 1.57-1.501.861-.629 1.759-.978 2.689-1.048 1.118-.093 1.921.173 2.41.8.418.51.605 1.241.559 2.191-.024.65-.181 1.595-.472 2.836-.292 1.241-.436 1.953-.436 2.139-.024.488.023.824.139 1.009.117.186.395.278.838.278.326 0 .907-.344 1.746-1.034.698-.573 1.269-1.113 1.712-1.619l.664.872'
-    />
-    <path
-      className='logo'
-      d='M52.295 10.654c.022-.625-.233-.938-.767-.938-.698 0-1.407.481-2.127 1.442-.721.961-1.093 1.882-1.116 2.762-.013 0-.013.151 0 .452a7.86 7.86 0 0 0 2.964-1.911c.674-.741 1.022-1.344 1.046-1.807zm7.927 3.646c-.675 1.117-2.002 2.232-3.981 3.348-2.467 1.418-4.971 2.126-7.508 2.126-1.885 0-3.237-.628-4.051-1.885-.582-.861-.861-1.885-.838-3.072.023-1.885.862-3.677 2.515-5.377 1.815-1.862 3.957-2.794 6.425-2.794 2.282 0 3.492.93 3.632 2.787.093 1.184-.559 2.404-1.956 3.658-1.49 1.371-3.365 2.241-5.622 2.612.418.581 1.046.871 1.885.871 1.676 0 3.504-.426 5.483-1.279 1.42-.599 2.538-1.221 3.353-1.866l.663.871'
-    />
-    <path
-      className='logo'
-      d='M65.755 11.828a4.407 4.407 0 0 0-.262-1.732c-.198-.524-.484-.788-.855-.788-1.188 0-2.166.642-2.933 1.925-.653 1.05-1.003 2.17-1.048 3.358-.024.584.081 1.098.314 1.54.255.514.616.77 1.083.77 1.047 0 1.944-.617 2.689-1.854.628-1.027.965-2.1 1.012-3.219zm3.946.132c-.093 2.139-.884 3.987-2.374 5.544-1.49 1.557-3.342 2.336-5.553 2.336-1.839 0-3.236-.593-4.19-1.779-.698-.883-1.083-1.987-1.152-3.311-.118-2 .604-3.836 2.165-5.51 1.676-1.859 3.782-2.789 6.32-2.789 1.629 0 2.863.547 3.702 1.639.792 1 1.152 2.29 1.082 3.87'
-    />
-  </Svg>
-)
-
-const FlickrBase = props => (
-  <Svg
-    xmlns='http://www.w3.org/2000/svg'
-    overflow='visible'
-    version='1'
-    height={57}
-    viewBox='0 0 564.893 155.72'
-    {...props}
-  >
-    <path
-      className='logo'
-      d='M22.622 73.44H0V45.588h23.262v-7.071c0-34.262 15.579-46.199 46.31-46.199 8.323 0 14.725 1.326 19.632 2.212l-2.346 27.408c-3.2-1.105-6.402-1.769-12.164-1.769-9.816 0-13.658 7.073-13.658 18.348v7.072h27.316V73.44H61.036v87.307H22.622V73.44zM108.41-5.029h38.413v165.776H108.41zM179.688-2.377h38.414V26.8h-38.414zM179.688 45.588h38.414v115.157h-38.414zM333.557 78.52c-7.043-4.417-14.085-6.408-23.262-6.408-16.646 0-29.45 11.936-29.45 31.83 0 18.347 14.939 30.28 31.584 30.28 8.751 0 17.502-1.987 23.903-5.524l.853 29.618c-9.604 3.318-21.127 5.086-31.37 5.086-36.709 0-64.664-22.104-64.664-60.122 0-38.24 27.955-60.342 64.663-60.342 11.525 0 21.767 1.99 30.518 6.41l-2.775 29.173zM353.193-5.029h38.412v99.026h.428l32.01-48.41h42.041l-38.626 52.827 41.828 62.333h-46.524l-30.73-57.026h-.427v57.026h-38.412V-5.03z'
-    />
-    <path
-      className='logo'
-      d='M562.97 77.416c-4.268-1.324-8.537-1.324-13.017-1.324-17.925 0-27.957 13.481-27.957 36.028v48.627h-38.412V45.587h34.997v21.22h.428c6.614-14.59 16.218-23.872 32.865-23.872 4.478 0 9.174.663 13.019 1.325l-1.923 33.156z'
-    />
-  </Svg>
-)
-
-const Flickr = styled(FlickrBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo:nth-child(1) {
-        fill: #0063db;
-      }
-      .logo:nth-child(2) {
-        fill: #ff0084;
-      }
-    `
-  })};
-`
-
-const Vimeo = styled(VimeoBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #1ab7ea;
-      }
-    `
-  })};
-`
-
-const RedditBase = props => (
-  <Svg
-    height={71.5}
-    viewBox='0 0 487 168'
-    fill='black'
-    xmlns='http://www.w3.org/2000/svg'
-    {...props}
-  >
-    <g fillRule='evenodd'>
-      <g transform='translate(201 40)'>
-        <circle className='logo' cx='238.4' cy='17.7' r='10.6' />
-        <path
-          d='M87 62.9c5.2 0 8.2-3.9 8-7.7-.1-2-.3-3.3-.5-4.3-3.2-13.8-14.4-24-27.7-24-15.8 0-28.6 14.3-28.6 31.9 0 17.6 12.8 31.9 28.6 31.9 9.9 0 17-3.6 22.1-9.3 2.5-2.8 2-7.1-1-9.2-2.5-1.7-5.6-1.1-7.9.8-2.2 1.9-6.4 5.7-13.2 5.7-8 0-14.8-6.9-16.3-16H87v.2zM66.7 39c7.2 0 13.4 5.6 15.7 13.3H51C53.3 44.5 59.5 39 66.7 39z'
-          fillRule='nonzero'
-        />
-        <path d='M38.8 33.2c0-3-2.2-5.4-5-5.9-8.7-1.3-16.9 1.9-21.6 7.8v-.8c0-3.7-3-6-6-6-3.3 0-6 2.7-6 6v49c0 3.2 2.4 6 5.6 6.2 3.5.2 6.4-2.5 6.4-6V58.2c0-11.2 8.4-20.4 20-19.1h1.2c3-.2 5.4-2.8 5.4-5.9zM244.4 39.6c0-3.3-2.7-6-6-6s-6 2.7-6 6v43.9c0 3.3 2.7 6 6 6s6-2.7 6-6V39.6z' />
-        <path
-          d='M155.2 6.4c0-3.3-2.7-6-6-6s-6 2.7-6 6v26.7C139.1 29 134 27 127.9 27c-15.8 0-28.6 14.3-28.6 31.9 0 17.6 12.8 31.9 28.6 31.9 6.1 0 11.3-2.1 15.4-6.2.5 2.8 3 4.9 5.9 4.9 3.3 0 6-2.7 6-6V6.4zM128 78.9c-9.2 0-16.6-8.9-16.6-20 0-11 7.4-20 16.6-20 9.2 0 16.6 8.9 16.6 20s-7.5 20-16.6 20zM217.8 6.4c0-3.3-2.7-6-6-6s-6 2.7-6 6v26.7c-4.1-4.1-9.2-6.1-15.3-6.1-15.8 0-28.6 14.3-28.6 31.9 0 17.6 12.8 31.9 28.6 31.9 6.1 0 11.3-2.1 15.4-6.2.5 2.8 3 4.9 5.9 4.9 3.3 0 6-2.7 6-6V6.4zm-27.3 72.5c-9.2 0-16.6-8.9-16.6-20 0-11 7.4-20 16.6-20 9.2 0 16.6 8.9 16.6 20s-7.4 20-16.6 20z'
-          fillRule='nonzero'
-        />
-        <path d='M274.8 83.6V39h5.3c2.8 0 5.3-2.1 5.5-5 .2-3.1-2.3-5.7-5.3-5.7h-5.5v-8.4c0-3.2-2.4-6-5.6-6.2-3.5-.2-6.4 2.5-6.4 6v8.7h-5c-2.8 0-5.3 2.1-5.5 5-.2 3.1 2.3 5.7 5.3 5.7h5.1v44.6c0 3.3 2.7 6 6 6 3.5-.2 6.1-2.8 6.1-6.1z' />
-      </g>
-      <g>
-        <circle className='logo' cx='83.8' cy='83.8' r='83.8' />
-        <path
-          d='M139.7 83.8c0-6.8-5.5-12.2-12.2-12.2-3.3 0-6.3 1.3-8.5 3.4-8.4-6-19.9-9.9-32.7-10.4l5.6-26.2 18.2 3.9c.2 4.6 4 8.3 8.7 8.3 4.8 0 8.7-3.9 8.7-8.7 0-4.8-3.9-8.7-8.7-8.7-3.4 0-6.4 2-7.8 4.9l-20.3-4.3c-.6-.1-1.2 0-1.6.3-.5.3-.8.8-.9 1.4L82 64.7c-13 .4-24.7 4.3-33.2 10.4-2.2-2.1-5.2-3.4-8.5-3.4-6.8 0-12.2 5.5-12.2 12.2 0 5 3 9.2 7.2 11.1-.2 1.2-.3 2.4-.3 3.7 0 18.8 21.9 34.1 48.9 34.1s48.9-15.2 48.9-34.1c0-1.2-.1-2.5-.3-3.7 4.2-1.9 7.2-6.2 7.2-11.2zm-83.8 8.7c0-4.8 3.9-8.7 8.7-8.7 4.8 0 8.7 3.9 8.7 8.7 0 4.8-3.9 8.7-8.7 8.7-4.8.1-8.7-3.9-8.7-8.7zm48.7 23.1c-6 6-17.4 6.4-20.7 6.4-3.4 0-14.8-.5-20.7-6.4-.9-.9-.9-2.3 0-3.2.9-.9 2.3-.9 3.2 0 3.8 3.8 11.8 5.1 17.5 5.1 5.7 0 13.8-1.3 17.5-5.1.9-.9 2.3-.9 3.2 0 .8.9.8 2.3 0 3.2zm-1.6-14.3c-4.8 0-8.7-3.9-8.7-8.7 0-4.8 3.9-8.7 8.7-8.7 4.8 0 8.7 3.9 8.7 8.7 0 4.7-3.9 8.7-8.7 8.7z'
-          fill='white'
-          fillRule='nonzero'
-        />
-      </g>
-    </g>
-  </Svg>
-)
-
-const Reddit = styled(RedditBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #ff4500;
-      }
-    `
-  })};
-`
-
-const NYTimes = props => (
-  <Svg height={27.3} viewBox='0 0 49.072 6.46' {...props}>
-    <path d='M3.683.662c0-.527-.495-.68-.908-.66v.082c.248.006.438.089.438.254 0 .114-.082.26-.33.26-.197 0-.578-.114-.883-.216C1.664.268 1.346.16 1.08.16c-.521 0-.915.388-.915.864 0 .387.292.527.413.578l.038-.064c-.076-.05-.152-.108-.152-.26 0-.115.101-.299.368-.299.241 0 .559.108.978.229.362.101.762.197.978.222v.825l-.407.337v.032l.407.356v1.136c-.21.121-.451.153-.66.153-.388 0-.756-.102-1.049-.426l1.093-.527V1.494l-1.34.597c.14-.356.419-.59.705-.762l-.02-.064C.725 1.481 0 2.211 0 3.113c0 1.08.876 1.88 1.873 1.88 1.08 0 1.766-.845 1.753-1.74l-.064-.006c-.158.343-.387.667-.685.838V2.98l.431-.343-.006-.038-.425-.343V1.43c.394 0 .806-.267.806-.769zM1.365 3.615l-.324.158c-.184-.24-.304-.565-.304-1.016 0-.184.019-.413.076-.571l.559-.248zm2.826.59l-.356.267.045.05.171-.12.597.546.8-.533-.038-.058-.21.134-.26-.26V2.4l.216-.158.457.368v1.62c0 1.01-.222 1.174-.679 1.34v.07c.756.031 1.435-.223 1.435-1.518v-1.74l.241-.191-.044-.057-.21.165-.654-.565-.762.571V.11h-.057l-.946.635v.058c.095.057.254.095.254.413zm4.877-.285l-.654.508-.68-.534v-.318l1.264-.844v-.038l-.629-.966-1.378.737v1.753l-.273.21.045.056.247-.196.902.654 1.207-.953zm-1.334-.44V2.173l.045-.032.584.921zM14.173.42a.748.748 0 0 0-.05-.248h-.064c-.083.223-.185.343-.445.343-.229 0-.406-.14-.502-.241 0 .006-.762.882-.762.882l.064.058.216-.248c.165.127.285.248.654.254v2.222L11.722.738c-.12-.203-.324-.508-.699-.508-.419 0-.806.362-.755.972h.076c.019-.165.101-.343.286-.343.146 0 .279.14.349.266v.87c-.483 0-.8.223-.8.616 0 .204.101.546.438.61l.006-.057c-.045-.045-.09-.095-.09-.184 0-.146.109-.242.305-.242.032 0 .121.006.14.013v1.13c-.565.006-1.003.318-1.003.851 0 .508.438.75.895.724l.006-.063c-.286-.032-.438-.165-.438-.35 0-.228.165-.348.375-.348.216 0 .406.133.54.285l.768-.857-.064-.057-.197.222c-.292-.26-.457-.356-.793-.406V1.208l2.152 3.753h.153v-3.74c.413-.026.8-.337.8-.8zm1.924 3.499l-.654.508-.68-.534v-.317l1.264-.845v-.038l-.629-.965-1.378.737v1.752l-.273.21.045.057.248-.197.901.654 1.207-.952-.051-.07zm-1.334-.439V2.173l.045-.032.584.921zm5.702-1.46l-.196.146-.508-.432-.59.534.24.228V4.51l-.647-.413V2.46l.21-.146-.623-.578-.578.534.241.228v1.911l-.082.051-.565-.4V2.453c0-.369-.178-.477-.413-.623-.197-.12-.286-.222-.286-.406 0-.152.153-.241.229-.286-.006-.006-.006-.063-.006-.063-.203-.006-.769.216-.781.73-.006.267.12.38.254.508.133.127.254.248.254.47v1.543l-.286.222.045.058.26-.204.705.553.666-.451.737.45 1.416-.825V2.34l.35-.255-.045-.063zM25.412.554l-.267.241-.596-.52-.87.647v-.61h-.09l.02 4.325c-.09-.013-.311-.063-.515-.095L23.075.93a.633.633 0 0 0-.666-.635c-.49 0-.807.374-.807.749h.076c.026-.159.108-.299.267-.299.165 0 .305.102.305.458v1.041c-.49.02-.769.292-.769.629 0 .21.096.533.439.546l.012-.064c-.114-.05-.14-.12-.14-.184 0-.152.147-.215.35-.215h.095v1.65c-.387.134-.565.42-.565.744 0 .444.356.78.876.78.369 0 .686-.063 1.023-.14.273-.056.603-.12.762-.12.203 0 .292.095.292.242 0 .183-.083.254-.19.285l.012.064c.438-.083.692-.343.692-.737 0-.406-.387-.648-.82-.648-.221 0-.66.07-.99.147-.368.082-.743.12-.857.12-.184 0-.413-.089-.413-.336 0-.223.184-.4.642-.4.247 0 .527.038.831.107.324.07.61.16.89.16.387 0 .736-.14.736-.706V.877l.311-.266zM24.32 2.199a.447.447 0 0 1-.318.152c-.146 0-.254-.089-.323-.152V1.01l.273-.184.368.337zm0 .793a.463.463 0 0 0-.318-.133.527.527 0 0 0-.323.133v-.698c.057.05.184.14.323.14.14 0 .254-.077.318-.14zm0 1.245c0 .216-.146.425-.426.425a2.44 2.44 0 0 1-.215-.012V3.088a.499.499 0 0 1 .323-.147c.134 0 .242.077.318.147zm3.645-1.893l-.864-.616-1.295.744v1.74l-.267.216.038.056.21-.17.857.628 1.32-.807zm-1.416 1.683V2.122l.673.483V4.51zM30.51 1.8c0 .006-.05 0-.05 0a.38.38 0 0 1-.236.107.421.421 0 0 1-.298-.145h-.05l-.452.495-.463-.495-.788.545.038.064.21-.14.273.292v1.683l-.356.267.045.05.171-.12.642.546.832-.553-.038-.057-.235.14-.312-.28V2.325c.134.146.293.273.477.273.355 0 .552-.393.59-.8zm3.207 2.559l-.908.597-1.232-1.861.87-1.352.057-.006c.102.101.28.21.445.21a.467.467 0 0 0 .387-.21s.05.006.05 0c-.03.54-.393.844-.666.844-.28 0-.413-.127-.559-.216l-.082.121 1.333 1.975.267-.165zm-2.934-.153l-.355.267.044.05.171-.12.597.546.8-.533-.044-.058-.203.134-.26-.26V.11h-.058l-.965.64v.058c.095.057.273.083.273.406zm7.36-3.543c0-.527-.496-.68-.908-.66v.082c.247.006.438.089.438.254 0 .114-.083.26-.33.26-.197 0-.578-.114-.883-.216-.337-.114-.654-.222-.92-.222-.522 0-.915.388-.915.864 0 .387.292.527.412.578l.039-.064c-.077-.05-.153-.107-.153-.26 0-.114.102-.299.368-.299.242 0 .56.108.978.229.362.102.762.197.978.223v.825l-.406.337v.031l.406.356v1.137c-.21.12-.45.152-.66.152-.387 0-.756-.102-1.048-.425l1.092-.527V1.494l-1.34.597c.14-.356.42-.59.705-.762l-.019-.064c-.793.216-1.517.946-1.517 1.848 0 1.08.876 1.88 1.873 1.88a1.71 1.71 0 0 0 1.753-1.74l-.064-.006c-.159.343-.387.667-.686.838V2.98l.432-.343-.006-.038-.425-.343V1.43c.393 0 .806-.267.806-.769zm-2.318 2.953l-.324.158c-.184-.24-.305-.565-.305-1.016 0-.184.02-.413.076-.571l.56-.248-.007 1.677zM39.058.414L39.02.408l-.533.45v.039l.45.508h.045l.533-.445-.005-.03-.45-.515zm.813 3.944l-.203.133-.26-.26V2.389l.253-.19-.05-.064-.197.152-.464-.552-.762.534.045.07.19-.128.235.292V4.23l-.343.26.038.051.178-.127.584.534.8-.534zm4.445-.032l-.19.127-.286-.26v-1.81l.273-.203-.045-.058-.235.178-.61-.565-.787.559-.603-.56-.775.553-.476-.552-.762.533.038.07.19-.127.26.286v1.727l-.209.21.61.514.59-.534-.24-.228V2.383l.228-.153.413.381v1.613l-.204.21.616.514.578-.533-.241-.229V2.37l.203-.133.438.374V4.2l-.178.184.623.565.819-.559zm2.318-.407l-.654.509-.68-.534v-.317l1.264-.845v-.038l-.629-.965-1.378.737v1.81l.92.666 1.207-.953zM45.3 3.481V2.173l.045-.032.584.921zm3.772-.248l-.514-.388c.343-.298.476-.699.476-.971a1.95 1.95 0 0 0-.013-.172h-.063c-.051.14-.153.26-.362.26-.203 0-.35-.114-.477-.254l-1.2.655v.971l.445.35c-.451.4-.534.679-.534.882 0 .267.121.458.35.546l.031-.057c-.057-.05-.108-.083-.108-.203 0-.09.108-.216.312-.216.26 0 .419.177.501.273 0-.006 1.156-.692 1.156-.692zm-.286-.8c-.178.317-.59.628-.819.806l-.292-.229v-.946c.114.254.394.47.698.47a.796.796 0 0 0 .413-.102zm-.47 2.12c-.133-.298-.444-.514-.762-.514-.076 0-.298 0-.501.127.12-.203.476-.584.933-.857l.33.254z' />
-  </Svg>
-)
-
-const TheGuardian = props => (
-  <Svg height={68} viewBox='0 0 295 97' {...props}>
-    <path d='M66.9 51.7c1-.5 2.1-1.1 3.1-1.6.4-.2 1.8-.7 2-1.1.1-.2 0-.8 0-1.1V9.5c0-.1.1-.8 0-.9-.1-.1-.9 0-1 0h-2.2c-.8 0-.7.1-1.2.7-1.9 2.5-3.8 5-5.7 7.6-.9 1.2-1.8 2.3-2.6 3.5-.6.8-.6.9-1.7.9l.3-6.6c.1-1.6.1-3.2.2-4.8 0-.6.1-1.2.1-1.8 0-.2-.1-.8 0-.9.2-.3 2-.1 2.5-.1h37.4c.2 0 1.1-.1 1.3 0 .2.1.1.7.1 1 .1 1.3.1 2.7.2 4 .1 2.9.2 5.7.4 8.6v.6c-.5 0-.9.1-1.3-.2-.4-.3-.7-.9-1-1.3-1.2-1.6-2.4-3.3-3.7-4.9l-3.6-4.8c-.2-.3-.5-1-.9-1.2-.3-.2-.9-.1-1.3-.1h-1.6c-.2 0-1-.1-1.2 0-.1.1 0 .1-.1.3-.2.8 0 1.8 0 2.6v34.8c0 .7-.2 1.7 0 2.4s1.8 1.2 2.5 1.5c.7.4 1.9.7 2.5 1.3.2.2.4 1.2.2 1.5-.1.1-.6 0-.7 0h-22c-.2 0-.9.1-1 0-.1-.4 0-1.2 0-1.5zm37.1-1.8V7.5c0-.6.2-1.5 0-2.1-.2-.7-1.4-.8-2.1-1.1-.4-.1-1.7-.4-1.9-.8v-.8h.3c.4-.2 1.2-.2 1.7-.3.9-.2 1.8-.3 2.6-.5 2-.4 4.1-.8 6.1-1.1 1.2-.2 2.4-.5 3.6-.7.3 0 1.5-.2 1.6 0 .1.1 0 .5 0 .6v20.7c3.8-3.8 10.4-5.9 15.7-4.6 4.6 1.1 6.1 5.4 6.1 9.8.1 5.6 0 11.1 0 16.7v4.5c0 .5-.2 1.7 0 2.2.1.2.5.3.7.5.7.4 1.5.8 2.2 1.2.1.1.4.2.4.3.1.1.1 1 0 1.1-.1.2-.1.1-.4.1-1.1.2-2.3 0-3.3 0h-14.4c-.1 0-.6.1-.7 0-.1-.1-.1-1.2 0-1.3.3-.4 1.5-.8 1.9-1.1.3-.2 1.2-.5 1.4-.8.2-.3 0-1.4 0-1.7v-4.1c0-5.8.1-11.7 0-17.5 0-1.7-.3-3.7-2-4.6-.9-.5-2.1-.6-3.1-.6-.8 0-1.6.2-2.3.4-.6.2-1.7.6-2.1 1.1-.3.4-.1 1-.1 1.5v23.6c0 .5-.2 1.5 0 1.9.1.2.5.4.8.5.7.4 1.6.8 2.3 1.2.1.1.3.1.3.2.1.1.1 1 0 1.1-.1.1-.2.1-.4.1-1.1.1-2.3 0-3.4 0h-14.4c-.1 0-.6.1-.7 0-.1-.1 0-1.1.1-1.3.3-.4 1.4-.8 1.8-1 .7-.3 1.2-.6 1.7-.9zm47.5-13.8c.2 4.4 1.6 9.3 5.6 11.7 2.1 1.2 4.7 1.6 7.1 1.4 2.9-.3 5.3-1.7 7.8-3 0 .4.1 1.1 0 1.5-.2.4-.7.9-1 1.2-.4.4-.9.9-1.4 1.3-1.7 1.4-3.6 2.4-5.6 3.1-3.8 1.3-8.2 1.2-12.1.2-3.1-.8-5.9-2.3-8-4.7-2.1-2.4-3.3-5.5-3.8-8.7-1.1-6.9.2-14.8 5.9-19.4 5-4 13-4.8 18.9-2.4 7.2 2.9 8.4 11.1 8.4 17.9h-19.1c-.9-.1-1.8-.1-2.7-.1zm-.2-1.7c2-.1 4-.2 6.1-.4 1.2-.1 2.5-.1 3.7-.2.2 0 .7 0 .9-.1.3-.2.1-2.2.1-2.6-.1-3.6-.2-8-2.3-11.1-1.5-2.2-3.8-1.8-5.2.2-2.9 3.9-3.3 9.7-3.3 14.2zM-.1 70.9c0-5.6 1.1-11.4 4.2-16.1 3-4.6 7.8-7.6 13-9.1 6.3-1.8 13.3-1.6 19.8-.4 1.4.3 2.8.6 4.1 1 .2.1 1.1.3 1.2.5v.3c.3 1.2.1 2.7.1 3.9.1 3.1.1 6.3.2 9.4-.4 0-1.1.1-1.4 0-.2-.1-.4-.6-.5-.8-.8-1.2-1.6-2.4-2.3-3.7-1.7-2.7-3.5-5.4-5.2-8.1-.9-1.5-3.4-1.5-5-1.5-1.8 0-3.6.3-5.1 1.2-3.4 2.1-5 6.2-5.9 9.9-1.4 5.9-1.3 12.1-1 18.1.3 4.6.7 9.5 2.5 13.7 1.1 2.6 3 4.9 5.9 5.5.9.2 1.8.2 2.8.1.6-.1 2.7-.2 3-.8.1-.2 0-.7 0-.9V76.7c0-.7.2-1-.4-1.4-.7-.4-1.5-.8-2.2-1.3-.5-.3-1.5-.6-1.8-1-.2-.2-.3-1.3-.1-1.5.1-.2.1-.1.4-.1.9-.2 2.1 0 3 0h15.5c1 0 2.1-.1 3.1 0h.4c.1.1.1 1.5 0 1.7-.2.3-1.2.7-1.6.9-.7.4-1.5.8-2.2 1.3-.8.4-.7.4-.7 1.3v16.3c0 .2.1.7 0 .9-.2.5-2.3.8-2.9 1-1.7.5-3.5.9-5.3 1.2-3.5.6-7 .8-10.5.7-5.6-.2-11.5-1.4-16.1-4.8-4-2.9-6.6-7.3-7.8-12.1C.2 77-.1 73.9-.1 70.9zM46 61.8c0-.2-.1-1 0-1.1.1-.1.2 0 .3-.1.6-.3 1.4-.3 2.1-.4 2.3-.4 4.6-.8 7-1.2 1.8-.3 3.6-.7 5.3-.9.5-.1 1.1-.1 1.6.1.5.1.3 0 .4.4.2 1.2 0 2.6 0 3.8v20.9c0 2.5-.7 7.1 1.7 8.6 1.2.8 2.9.8 4.3.5 1.1-.2 3.2-1 3.6-2.2.1-.3 0-.7 0-1V64.7c0-.2.1-.8 0-1-.3-.3-1.3-.6-1.6-.7-.6-.2-1.1-.5-1.7-.7-.2-.1-.9-.2-1-.4-.1-.1-.1-1 0-1.1.1-.1.1 0 .2-.1.5-.3 1.4-.2 1.9-.3 4.1-.7 8.3-1.6 12.5-2.2.5-.1 1-.1 1.5 0s.3-.1.4.4c.2 1 0 2.4 0 3.5v28.8c0 .3-.1.8 0 1.2.1.3 0 .2.5.4 1 .4 2.1.9 3.1 1.3.3.1.4.1.5.3.1.2.1.9 0 .9-.2.2-1.2.1-1.4.2-2.2.3-4.4.5-6.7.8-2 .2-4 .5-5.9.7-.4.1-.8.1-1.2 0-.2 0-.8 0-.9-.1-.1-.2 0-1 0-1.2v-1.7c0-.3.2-1.1 0-1.4-.2-.3-.6.1-.9.3-.8.7-1.7 1.4-2.6 1.9-3.3 2-7.4 2.9-11.2 2.1-2.8-.6-5.2-2.2-6.4-4.9-.9-1.9-1.1-4-1.1-6v-21c0-.2.1-.8 0-1-.2-.3-1.4-.6-1.7-.8-1-.4-1.8-.7-2.6-1.1zm94.3-3.8c.2 0 1.1 0 1.2.2.1.1 0 .8 0 .9v8.8c0 .3-.1.8 0 1.1v.2c.4.4.5-.9.6-1 .9-3.7 2.6-8 6.4-9.5.7-.3 1.4-.4 2.1-.5.6 0 2.3-.2 2.6.3.1.2 0 .7 0 .9v10.2c-2.4-.5-5.1-.4-7.4.1-1.1.2-2.1.5-3.1 1-.5.2-.6.1-.7.5-.1.4 0 1.1 0 1.5V92c0 .2-.1.5 0 .7.2.3.9.5 1.2.6.5.3 1.1.6 1.6.9.2.1.5.2.6.4.1.1.1 1.3 0 1.4-.1.1-.7 0-.9 0h-16.6c-.5 0-1.4.2-1.8 0h-.3c-.1-.1 0-1.2 0-1.3.2-.6 1.5-1 2.1-1.3.4-.2 1.2-.5 1.4-.9.1-.2 0-.8 0-1V63.7c0-.1.1-.6 0-.7-.1-.1-.6-.2-.8-.2-.6-.2-1.2-.4-1.8-.5-.6-.2-2.1-.3-1.6-1.3.1-.3.5-.2.9-.3 4.4-.8 8.7-1.6 13.1-2.4.4-.2.8-.2 1.2-.3zm37.7.9V48.6c0-.3.1-.9 0-1.2-.1-.2-.5-.3-.8-.4-.8-.3-1.6-.6-2.5-.9-.6-.2-.9-.1-.9-.8 0-.6-.1-.4.4-.5 1.1-.4 2.4-.4 3.5-.6 3.3-.6 6.5-1.2 9.8-1.8.7-.1 1.4-.3 2.1-.2.8.1.9-.1.9.6.1 1.6 0 3.3 0 4.9v43.8c0 .4-.1.4.2.6.6.4 1.6.6 2.2.8.4.1 1.5.3 1.7.6.1.1.1 1.2 0 1.3-.1.1-.7.1-.8.1-1.4.2-2.7.4-4.1.5-3.1.4-6.2.8-9.4 1.2-.5.1-1.7.3-1.9-.1-.2-.4 0-1.7 0-2.1 0-.3.2-1.6 0-1.8-.2-.2-2.4 1.8-2.7 2-1.2.8-2.5 1.4-3.9 1.8-2.2.6-4.7.6-6.9.1-9.6-2.2-11.6-13.6-10.8-21.9.4-4.4 1.8-8.8 4.8-12 2.8-2.9 6.6-4.4 10.5-4.7 3-.3 5.9-.1 8.6 1zm0 31.9V61.6c0-.6-.1-.7-.8-1.1-1.4-.8-3.2-1.1-4.7-.5-2.3.9-3.4 3.6-4 5.8-1.2 4-1.3 8.3-1.2 12.5.1 3.6.2 8 2.2 11.2 1.8 2.8 5.9 3.5 8.5 1.3zM211.1 58c.5.1 1.1-.1 1.3.4.1.3 0 1 0 1.3V91c0 .4-.1 1 0 1.4.1.4.5.5.9.7.6.3 1.1.6 1.7.9.2.1.8.3.9.6 0 .1.1 1.3 0 1.4h-.3c-.6.1-1.4 0-2.1 0h-15.2c-.4 0-1.5.2-1.8 0-.2-.1-.3-1.3-.1-1.5.2-.3.9-.5 1.2-.6.6-.3 1.2-.7 1.8-1 .5-.3.4-.2.4-.6.2-1.5 0-3.2 0-4.7V65.3c0-.4.1-1.1 0-1.5-.1-.4-.6-.4-1-.6-.7-.3-1.3-.5-2-.8-.2-.1-1-.2-1.2-.5-.1-.1-.1-1.1 0-1.2.1-.1.3 0 .4-.1.8-.3 1.7-.3 2.6-.5 1.3-.2 2.5-.5 3.8-.7 2.5-.4 4.9-.9 7.4-1.3.5.1.9 0 1.3-.1zm1.4-9.3c-.2 6-8.3 8.6-12 3.9-1.7-2.2-1.8-5.2-.3-7.5s4.4-3.3 7.1-2.7c3 .6 5.2 3.2 5.2 6.3zm43.8 43.9V63.9c0-.1.1-.6 0-.7-.1-.2-.7-.3-.9-.3l-1.8-.6c-.3-.1-1.1-.2-1.3-.5-.2-.2-.1-.7-.1-1 0-.6.2-.5.8-.6l6.6-1.2 6.6-1.2c.4-.1.7-.2 1.1-.2.2 0 1.2 0 1.4.1.1.1 0 .7 0 .9v1.9c0 .3-.2 1.2 0 1.5.2.4.6 0 .9-.3.9-.7 1.8-1.3 2.8-1.8 3.6-2 7.9-2.9 12-2.3 2.8.4 5.1 1.9 6.1 4.6.7 1.9.8 3.9.8 5.9v22.8c0 .4-.1 1.1 0 1.5.1.4.4.5.9.7.5.3 1.1.6 1.6.9.2.1.9.3 1 .6.1.2.1 1.2 0 1.3-.1.2.1 0-.2.1-.4.2-1.3 0-1.8 0h-14.3c-.6 0-2.9.3-3.2-.1-.1-.1-.1-1.2 0-1.4.1-.2.7-.4.9-.5.6-.3 1.1-.6 1.7-.9.2-.1.8-.3.9-.5.1-.2 0-.6 0-.8V69.1c0-1.6.1-3.4-.9-4.8-.8-1.2-2.2-1.5-3.6-1.6-1.1 0-2.1.1-3.1.4-.5.2-2 .6-2.3 1.2-.1.2 0 .6 0 .8V91c0 .3-.1 1 0 1.3.1.4.7.5 1 .7.5.3 1.1.6 1.6.9.2.1.6.3.7.4.1.1.1 1.3 0 1.4 0 .1-.1 0-.2 0-.4.2-1.4 0-1.8 0h-16.8c-.5 0-.6.1-.7-.2-.1-.2-.1-1.1 0-1.3.1-.2.7-.4.9-.5 1-.1 1.8-.6 2.7-1.1zm-21.2-18.4c0-2.6.1-5.3-.1-8-.3-2.9-1.1-6.1-4.4-6.7-1-.2-2.8-.5-3.5.3-.7.7-1.2 1.7-1.8 2.5l-5.1 6.9c-.3.4-.6 1.2-1 1.4-.1.1-1.1.1-1.1 0-.1-.1 0-.8 0-1v-8.1c0-.3-.1-.8 0-1.1.2-.5 2.4-.8 2.9-1 2.8-.7 5.7-1.2 8.6-1.4 4.1-.3 8.7-.2 12.5 1.6 3 1.4 4.9 3.9 5.4 7.1.3 1.6.2 3.4.2 5v20c0 .3-.2 1.1 0 1.3.1.1.6.2.7.2.8.2 1.5.4 2.3.6.5.1.6.1.6.6s.1.5-.5.7c-.7.3-1.4.6-2.1.8-3.7 1.1-9.5 1.4-12.2-2-.2-.3-.8-1.6-1.1-1.7-.4-.1-1.6 1.5-2 1.8-1.6 1.4-3.7 2.2-5.8 2.5-3.9.5-8.1-.3-10.5-3.7-2-2.9-2.4-7.4-.9-10.6 2.7-6.1 10.7-6.7 16.5-7.7.7 0 1.5-.2 2.4-.3zm0 16.6V76c-3 .2-6.1.4-7.2 3.7-.8 2.2-.8 4.8-.5 7.1.2 1.6.6 3.6 2.1 4.6 1.6 1.1 4.3.9 5.6-.6zM108.5 74.2c0-2.6.1-5.3-.1-8-.3-2.9-1.1-6.1-4.4-6.7-1-.2-2.8-.5-3.5.3-.7.7-1.2 1.7-1.8 2.5l-5.1 6.9c-.3.4-.6 1.2-1 1.4-.1.1-1.1.1-1.1 0-.1-.1 0-.8 0-1v-8c0-.3-.1-.8 0-1.1.2-.5 2.3-.8 2.9-1 2.8-.7 5.7-1.2 8.5-1.4 4.1-.3 8.7-.2 12.5 1.6 3 1.4 4.9 3.9 5.4 7.1.3 1.6.2 3.4.2 5v20c0 .3-.2 1.1 0 1.3.1.1.6.2.7.2.8.2 1.5.4 2.3.6.5.1.6.1.6.6s.1.5-.5.7c-.7.3-1.4.6-2.1.8-3.7 1.1-9.5 1.4-12.2-2-.2-.3-.8-1.6-1.1-1.7-.4-.1-1.6 1.5-1.9 1.8-1.6 1.4-3.7 2.2-5.8 2.5-3.9.5-8.1-.3-10.5-3.7-2-2.9-2.4-7.4-.9-10.6 2.7-6.1 10.7-6.7 16.4-7.7.8-.1 1.6-.3 2.5-.4zm0 16.6V76c-3 .2-6.1.4-7.2 3.7-.8 2.2-.8 4.8-.5 7.1.2 1.6.6 3.6 2.1 4.6 1.6 1.1 4.3.9 5.6-.6z' />
-  </Svg>
-)
-
-const CNNBase = props => (
-  <Svg height={96} viewBox='0 0 1000 464.6231995' {...props}>
-    <path
-      className='logo'
-      d='M75.814 232.456c0 85.224 69.334 154.558 154.558 154.558h144.366c9.066 0 16.098-8.618 16.098-16.04V87.745c0-17.892 14.557-32.453 32.444-32.453 11.476 0 22.17 6.084 27.902 15.886.718 1.234 44.427 76.59 90.698 156.377 44.21 76.24 89.934 155.077 90.652 156.315 2.314 3.945 6.836 6.491 11.53 6.491 7.434 0 13.485-6.047 13.485-13.485V87.746c0-17.893 14.549-32.454 32.445-32.454 11.475 0 22.17 6.084 27.897 15.886.727 1.234 44.398 76.59 90.64 156.381 44.17 76.224 89.85 155.04 90.586 156.31 2.313 3.946 6.836 6.492 11.53 6.492 7.434 0 13.476-6.047 13.476-13.485V2.023h-56.903v223.131S764.115 47.691 760.775 41.984C745.725 16.257 717.183 0 686.807 0c-47.584 0-86.172 38.604-86.172 86.184v138.97s-103.1-177.463-106.442-183.17C479.14 16.257 450.6 0 420.223 0c-47.583 0-86.17 38.604-86.17 86.184V314.54c.029 8.31-6.218 15.554-15.313 15.57h-87.758c-53.93 0-97.65-43.72-97.65-97.654s43.72-97.655 97.65-97.655h73.508V77.898h-74.118c-85.224 0-154.558 69.334-154.558 154.558'
-    />
-    <path
-      className='logo'
-      d='M943.093 2.02v374.858c0 17.896-14.55 32.453-32.445 32.453-11.476 0-22.17-6.084-27.902-15.882-.718-1.233-44.398-76.593-90.631-156.38-44.178-76.22-89.847-155.045-90.595-156.311-2.305-3.946-6.836-6.496-11.529-6.496-7.434 0-13.477 6.047-13.477 13.481v289.135c0 17.896-14.557 32.453-32.45 32.453-11.475 0-22.16-6.084-27.9-15.882-.719-1.233-44.427-76.59-90.699-156.38-44.215-76.237-89.933-155.074-90.656-156.311-2.313-3.946-6.836-6.496-11.53-6.496-7.434 0-13.477 6.047-13.477 13.481v283.233c0 18.648-16.384 35.008-35.061 35.008H230.375c-95.682 0-173.53-77.84-173.53-173.53 0-95.682 77.848-173.526 173.53-173.526h74.114V2.021h-74.048C103.174 2.02 0 105.187 0 232.454c0 127.271 103.174 230.442 230.441 230.442h145.75c55.212.029 90.718-32.32 90.643-91.99V239.469s103.84 178.725 106.447 183.173c15.052 25.725 43.593 41.981 73.97 41.981 47.587 0 86.166-38.604 86.166-86.18V239.47S836.521 416.936 839.86 422.642c15.056 25.725 43.596 41.981 73.973 41.981 47.584 0 86.167-38.604 86.167-86.18V2.022h-56.907z'
-    />
-    <path
-      fill='#FFF'
-      d='M56.847 232.456c0 95.686 77.844 173.53 173.526 173.53h144.365c18.677 0 35.066-16.36 35.066-35.012V87.745c0-7.434 6.043-13.481 13.477-13.481 4.693 0 9.216 2.546 11.53 6.491.718 1.242 46.44 80.078 90.656 156.315 46.271 79.788 89.975 155.144 90.698 156.381 5.736 9.798 16.422 15.878 27.897 15.878 17.897 0 32.454-14.557 32.454-32.453V87.746c0-7.435 6.043-13.482 13.477-13.482 4.693 0 9.22 2.546 11.53 6.491.743 1.271 46.416 80.091 90.593 156.31a912010.963 912010.963 0 0 0 90.632 156.386c5.731 9.798 16.422 15.878 27.897 15.878 17.896 0 32.45-14.557 32.45-32.453V2.023h-18.973v374.853c0 7.438-6.043 13.485-13.477 13.485-4.693 0-9.216-2.546-11.53-6.491-.738-1.271-46.416-80.087-90.585-156.31-46.242-79.793-89.913-155.148-90.64-156.382-5.731-9.802-16.422-15.886-27.897-15.886-17.896 0-32.45 14.561-32.45 32.453v289.13c0 7.44-6.046 13.486-13.48 13.486-4.694 0-9.217-2.546-11.53-6.491l-90.657-156.315a856792.202 856792.202 0 0 0-90.698-156.377c-5.731-9.802-16.421-15.886-27.897-15.886-17.892 0-32.445 14.561-32.445 32.453v283.23c0 7.42-7.031 16.039-16.098 16.039H230.373c-85.225 0-154.558-69.334-154.558-154.558 0-85.224 69.333-154.558 154.558-154.558h74.114V58.926h-74.114c-95.682 0-173.526 77.844-173.526 173.53'
-    />
-  </Svg>
-)
-
-const CNN = styled(CNNBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #c00;
-      }
-    `
-  })};
-`
-
-const TheWashingtonPost = props => (
-  <Svg height={31.7} viewBox='0 0 122.227 18.729' {...props}>
-    <path d='M6.804 5.766v6.625c.895-.484 1.522-1.325 1.826-2.274l.072.036c-.197 2.363-1.809 4.726-4.315 4.726C1.862 14.88 0 13 0 9.991c0-2.256 1.468-3.67 3.348-4.942a3.283 3.283 0 0 0-.841-.107c-1.29 0-2.042.877-2.042 1.773H.358C.34 6.589.34 6.464.34 6.339c0-1.594.842-3.528 2.883-3.528 1.307 0 2.31 1.2 3.832 1.2.608 0 1.36-.287 1.665-.95h.053c-.017 1.165-.375 2.365-1.969 2.705zm.967 6.786c-.896.949-1.862 1.755-3.116 1.755-2.327 0-4.118-1.827-4.118-4.477 0-1.558.645-2.399 1.361-3.384h-.072C1.38 6.894.197 8.022.197 9.937c0 2.937 1.844 4.746 4.261 4.746 1.684 0 2.83-1.164 3.349-2.113zm.716-8.684c-.304.573-.949 1.307-2.256 1.307s-2.507-.896-3.563-.896c-.967 0-1.63.645-1.88 1.129l.018.036c.34-.34.859-.681 1.7-.681 1.397 0 2.292.86 3.707.86 1.504 0 2.184-.985 2.31-1.737zM3.509 5.103c-.913.95-1.755 1.988-1.755 4.011 0 1.11.376 2.256 1.218 2.972l.52-.269V7.02l-.484.25-.09-.178L5.425 5.73c-.68-.143-1.289-.448-1.916-.627zm3.098.699c-.126.017-.25.017-.394.017-.25 0-.483-.017-.716-.071v5.228L3.15 12.248c.484.358 1.074.572 1.826.572.61 0 1.146-.125 1.63-.34V5.802zm-2.578.913l-.34.197v4.798l.34-.179zm10.295.68v6.195c0 1.72-1.522 2.883-3.223 3.223l-.036-.071a2.371 2.371 0 0 0 1.325-2.15v-6.41l-.966-.876-.376.411v4.996l.537.483v.036l-1.45 1.576-1.665-1.522v-.072l.644-.68V4.62l1.898-2.023.036.018v4.798l1.558-1.683 1.629 1.47.304-.323.144.126zM9.311 13.25v-.036l.376-.376V4.315l-.376.395v7.95l-.573.572v.036l1.396 1.253.25-.286zm3.652-5.354L11.8 6.894l-.25.268 1.038.932v6.445c0 .43-.09.877-.25 1.128l.018.018c.447-.322.609-.806.609-1.432V7.896zm4.208 3.098v.913l1.737 1.36 1.002-1.073.144.143-2.328 2.488-2.363-1.897-.358.376-.144-.144.412-.43V8.2l3.384-2.47 1.773 2.865zm-1.325 1.576V8.022l-.376.268v4.459l2.22 1.808.287-.304zm1.36-5.515l-.035.018v3.67l1.594-1.163zm.466-.358l-.322.232 1.576 2.525.322-.233zm14.429 8.111l-2.489-2.149-1.898 2.149-2.578-2.22v-3.51h-.501c-.448 0-.68.287-.788.627h-.054a2.845 2.845 0 0 1-.09-.68c0-.466.126-1.684 1.433-1.684v-3.08c0-.59-.698-.823-.698-1.61 0-1.021.966-2.024 2.74-2.651l.07.054c-.644.376-.984.734-.984 1.504 0 1.182 1.146.877 1.146 2.74v.716l2.095-2.203 2.202 2.167 2.077-2.167 2.006 1.97v6.625zM25.71 12.3V8.523h-1.003c-.52 0-.752.34-.788.645l.018.018c.18-.197.358-.304.77-.304h.627v3.616l2.363 2.042.269-.323zm0-8.218c0-.734-.644-.985-.644-1.45 0-.645.25-1.182.644-1.684l-.036-.017c-.537.412-1.038 1.074-1.038 1.718 0 .645.698.95.698 1.522v3.17h.376zm3.778 1.79l-1.36-1.325-.717.734v6.338L29.04 13l.447-.502zm.573 6.392V5.623l-1.522-1.504-.287.286 1.433 1.379v6.678l2.417 2.095.287-.269zm3.474-6.625l-1.128-1.074-.645.662v6.375l1.755 1.54.018-.018zm.572-.25L32.8 4.119l-.268.286 1.2 1.146v7.484l.375-.358zm7.607 9.419l-1.7-1.54-1.522 1.54-1.863-1.648-.25.269-.144-.143.323-.34V8.147l3.366-2.435 1.934 1.504.34-.358.144.143-.394.412v4.96l.948.877.305-.322.143.125zm-4.583-1.988V7.986l-.376.268V13l1.719 1.54.286-.287zm2.882-4.709l-1.468-1.146-.053.036v5.21l1.217 1.057.304-.304zm.574 4.87V7.844l-1.522-1.236-.358.25L40.21 8.04v5.12l1.504 1.361.268-.286zm6.338-4.01h1.844v3.831l-2.847 2.006c-.376-.43-.913-.734-1.575-.734-.538 0-.968.179-1.54.662l-.09-.036 2.435-3.545h-1.504v-3.33l2.99-2.077c.359.286.627.447 1.093.447.358 0 .877-.071 1.271-.393l.072.035zm-2.005 4.583c-.52 0-.931.18-1.29.43v.036c.287-.125.573-.161.842-.161.412 0 1.056.197 1.486.698l.412-.304c-.34-.394-.752-.699-1.45-.699zm1.915-3.008h-1.003l-2.095 3.027.018.035c.52-.376 1.021-.626 1.63-.626.537 0 1.092.25 1.432.68l.018-.018zm-2.614 0V7.664l-.376.268v3.009h1.45l.287-.394h-1.36zm3.742-3.51c-.376.27-.59.413-1.217.413-.358 0-.788-.161-1.128-.52l-.036.018v2.615h.68l1.72-2.507zm-.859-.16c-.322 0-.734-.161-1.003-.52l-.447.305c.34.412.662.59 1.163.59.322 0 .788-.089 1.272-.501l.017-.053a2.76 2.76 0 0 1-1.002.179zm.304 3.08h-1.164l-.268.376h1.056v3.473l.376-.268zm7.538-2.56v6.194c0 1.72-1.522 2.883-3.223 3.223l-.035-.071a2.371 2.371 0 0 0 1.325-2.15v-6.41l-.967-.876-.376.411v4.996l.537.483v.036l-1.45 1.576-1.665-1.522v-.072l.644-.68V4.62l1.898-2.023.036.018v4.798l1.558-1.683 1.63 1.469.303-.323.144.126zm-5.013 5.854v-.036l.376-.376V4.315l-.376.394v7.95l-.573.573v.036l1.396 1.253.25-.286zm3.653-5.354l-1.164-1.002-.25.268 1.038.932v6.445c0 .43-.09.877-.251 1.128l.018.018c.448-.322.609-.806.609-1.432V7.896zm4.62 6.912l-1.54-1.415-.359.376-.143-.143.412-.448V8.004l-.752-.716-.287.322-.143-.143 1.611-1.755 1.432 1.29.359-.395.16.126-.447.483v5.229l.86.805.394-.412.143.144zm-1.057-1.755V7.717l-.931-.859-.251.287.806.77v5.317l1.432 1.29.269-.287zm.537-7.484l-1.468-1.343 1.343-1.432 1.468 1.342zm-.913-1.63l-.269.269L57.668 5.3l.268-.304zm9.003 10.869l-1.558-1.415V8.165l-.949-.86-.501.538v4.834l.555.52v.07l-1.468 1.54-1.665-1.521v-.054l.644-.68V8.04l-.841-.77-.34.376-.144-.143 1.63-1.773 1.629 1.45v.358l1.7-1.808 1.594 1.432.358-.376.162.143-.43.448v5.12l.806.753.394-.43.143.143zM61.05 13.25v-.036c0 .036.358-.358.358-.358V7.771l-1.02-.931-.251.268.895.824v4.727l-.555.573v.036l1.379 1.253.286-.304zm3.742-.143V7.86l-1.11-1.003-.286.304 1.02.913v5.21l1.343 1.254.286-.304zm6.034 4.924c-.59-.645-1.002-.914-1.557-.914-.591 0-1.272.233-1.97.61l-.054-.054 2.471-2.865-1.754-1.397-.34.359-.144-.126.394-.43V8.148l3.402-2.435 1.898 1.504.322-.358.143.143-.376.412v6.124c0 .93.52.895.52 1.593 0 .806-1.164 1.558-2.955 2.9zm-2.381-4.96V8.004l-.376.269v4.96l1.772 1.432.27-.323zm1.253 3.581c-.59 0-1.128.18-1.629.52v.035c.322-.125.68-.286 1.29-.286.554 0 1.02.304 1.485.841l.358-.286c-.447-.52-.985-.824-1.504-.824zm2.185-1.719c-.448-.286-.555-.59-.555-1.36v-.323l-2.973 3.438.018.017c.573-.34 1.093-.59 1.934-.59.716 0 1.2.286 1.755.752.34-.287.573-.573.573-1.02 0-.359-.305-.627-.752-.914zm-.555-6.822L69.84 6.948l-.036.035v5.497l1.147.895.376-.43zm1.092 6.572c-.466-.359-.52-.663-.52-1.433V7.807L70.395 6.59l-.34.233 1.468 1.164V13.5c0 .77.036.985.52 1.307.43.286.77.59.77 1.002 0 .108-.018.287-.018.287l.018.018a.63.63 0 0 0 .197-.466c0-.394-.197-.662-.591-.966zm4.297.125l-1.647-1.29-.34.377-.143-.144.393-.43V7.056h-1.038l-.036-.036.627-.931h.447v-1.2l1.898-2.023.036.036v3.187h1.432l.036.036-.627.931h-.841v5.425l.985.753.376-.412.143.143zM75.552 4.584l-.376.412v1.092h.376zm0 8.559V7.055h-.376v6.302l1.522 1.182.269-.286zm8.72-5.569v5.014l-3.026 2.238-2.274-1.72-.269.305-.143-.143.322-.358V8.004l3.17-2.274 2.112 1.647.34-.358.144.143zm-4.816 5.157V7.843l-.377.268v4.817l2.15 1.647.34-.25zm2.882-4.566l-1.504-1.182-.018.018v5.336l1.522 1.146zm.573-.322l-1.665-1.29-.34.251 1.63 1.253v5.569l.375-.286zm8.2 6.965l-1.557-1.415V8.165l-.949-.86-.501.538v4.834l.555.52v.07l-1.468 1.54-1.666-1.521v-.054l.645-.68V8.04l-.842-.77-.34.376-.143-.143 1.63-1.773 1.629 1.45v.358l1.7-1.808 1.594 1.432.358-.376.161.143-.43.448v5.12l.806.753.394-.43.143.143zm-4.726-1.558v-.036c0 .036.358-.358.358-.358V7.771l-1.02-.931-.251.268.895.824v4.727l-.555.573v.036l1.379 1.253.286-.304zm3.742-.143V7.86l-1.11-1.003-.287.304 1.02.913v5.21l1.344 1.254.286-.304zm12.046 1.701c-.501-.537-.68-.717-1.271-.878v3.51l-.036.017-.716-.644-2.077 1.916-.036-.018v-4.816c-.627.143-1.29.465-1.934 1.11l-.036-.018c.125-1.379.806-2.4 1.97-2.793V8.702h-.377c-.519 0-.805.269-1.056.752h-.054c-.053-.16-.09-.412-.09-.823 0-.753.556-1.54 1.47-1.54h.106V5.157l-.734-.663-.358.359-.125-.126 1.808-1.916 1.683 1.54v1.63l.394-.43v-2.06h.197v1.88l2.381-2.56 2.023 1.845v7.788zm-4.136-1.576c-.734.143-1.253.59-1.593 1.182l.018.036c.501-.43.966-.627 1.575-.77zm.573-5.138h-.985c-.68 0-.949.43-.966.93l.017.018c.215-.358.43-.483.86-.536h.698v9.758l.376-.359zm0-3.224l-.895-.823-.287.304.806.716v2.024h.376zm2.095 1.02l-.394.413V16.67l.394.358zm2.346-.41l-1.057-.968-1.092 1.182v6.303c.931.09 1.576.358 2.13.967l.019-.018zm-2.149 7.752v.466c.609.143.86.322 1.29.823l.411-.322c-.43-.537-.86-.86-1.701-.967zm2.722-8.039l-1.218-1.128-.287.304 1.129 1.039v8.325l.376-.286zm7.931 2.381v5.014l-3.026 2.238-2.274-1.72-.268.305-.144-.143.323-.358V8.004l3.17-2.274 2.112 1.647.34-.358.143.143zm-4.816 5.157V7.843l-.376.268v4.817l2.148 1.647.34-.25zm2.883-4.566l-1.504-1.182-.018.018v5.336l1.522 1.146zm.573-.322l-1.666-1.29-.34.251 1.63 1.253v5.569l.376-.286zm5.514 1.128h1.845v3.831l-2.847 2.006c-.376-.43-.913-.734-1.576-.734-.537 0-.967.179-1.54.662l-.09-.036 2.436-3.545h-1.504v-3.33l2.99-2.077c.358.286.626.447 1.092.447.358 0 .877-.071 1.271-.393l.072.035zm-2.005 4.583c-.52 0-.931.18-1.29.43v.036c.287-.125.574-.161.842-.161.412 0 1.057.197 1.486.698l.412-.304c-.34-.394-.752-.699-1.45-.699zm1.916-3.008h-1.003l-2.094 3.027.017.035c.52-.376 1.02-.626 1.63-.626.537 0 1.092.25 1.432.68l.018-.018v-3.098zm-2.614 0V7.664l-.376.268v3.009h1.45l.286-.394h-1.36zm3.742-3.51c-.376.27-.59.413-1.218.413-.358 0-.788-.161-1.128-.52l-.036.018v2.615h.68l1.72-2.507zm-.86-.16c-.322 0-.734-.161-1.002-.52l-.448.305c.34.412.663.59 1.164.59.322 0 .788-.089 1.271-.501l.018-.053c-.304.125-.68.179-1.003.179zm.305 3.08h-1.164l-.268.376h1.056v3.473l.376-.268zm4.333 4.852l-1.647-1.29-.34.377-.144-.144.394-.43V7.056h-1.039l-.035-.036.626-.931h.448v-1.2l1.898-2.023.036.036v3.187h1.432l.036.036-.627.931h-.841v5.425l.985.753.376-.412.143.143zm-1.164-10.224l-.376.412v1.092h.376zm0 8.559V7.055h-.376v6.302l1.522 1.182.268-.286z' />
-  </Svg>
-)
-
-const EtsyBase = props => (
-  <Svg aria-hidden='true' height={103.5} viewBox='0 0 48 24' {...props}>
-    <path
-      className='logo'
-      d='M6.547 3.125v6.008s2.117 0 3.25-.086c.891-.156 1.055-.242 1.219-1.133l.328-1.305h.969l-.164 2.852.086 2.922h-.977l-.242-1.141c-.242-.812-.57-.977-1.219-1.055-.812-.086-3.25-.086-3.25-.086v5.039c0 .969.492 1.383 1.625 1.383h3.414c1.055 0 2.109-.086 2.766-1.625l.883-1.953h.82c-.086.406-.492 3.984-.57 4.789 0 0-3.086-.078-4.383-.078H5.25l-3.492.078v-.883l1.133-.25c.82-.164 1.062-.406 1.062-1.055 0 0 .086-2.195.086-5.852 0-3.648-.086-5.844-.086-5.844 0-.727-.242-.891-1.062-1.055l-1.133-.24v-.891l3.414.07h6.5c1.297 0 3.484-.234 3.484-.234s-.078 1.375-.164 4.625h-.891l-.328-1.141c-.32-1.461-.805-2.188-1.703-2.188H6.961c-.414.001-.414.079-.414.329zm13.156.641h.977V7.18l3.336-.164-.164 1.547-3.25-.25v6.016c0 1.703.57 2.359 1.547 2.359.883 0 1.539-.492 1.781-.898l.484.57c-.484 1.133-1.859 1.703-3.164 1.703-1.617 0-2.93-.969-2.93-2.836V8.398h-1.938v-.812c1.626-.164 2.837-1.141 3.321-3.82zm6.992 10.476l.648 1.547c.242.648.812 1.305 2.109 1.305 1.383 0 1.953-.734 1.953-1.625 0-2.766-5.445-1.953-5.445-5.688 0-2.109 1.703-3.094 3.898-3.094.977 0 2.438.164 3.172.492-.164.812-.25 1.867-.25 2.68l-.805.078-.57-1.625c-.164-.398-.82-.727-1.625-.727-.977 0-1.953.406-1.953 1.461 0 2.516 5.609 1.953 5.609 5.688 0 2.117-1.867 3.25-4.148 3.25-1.703 0-3.414-.656-3.414-.656.164-.969.086-2.023 0-3.086h.821zm6.336 7.797c.242-.891.406-2.023.57-3.086l.891-.078.328 1.703c.078.406.32.734.969.734 1.055 0 2.438-.648 3.742-2.922-.578-1.383-2.281-5.844-3.828-9.258-.406-.898-.484-.977-1.047-1.141l-.414-.156v-.82l2.445.086 3-.164v.813l-.734.164c-.57.078-.805.398-.805.727 0 .086 0 .164.078.328.156.492 1.461 4.141 2.438 6.578.805-1.703 2.352-5.523 2.594-6.172.086-.328.164-.406.164-.648 0-.414-.242-.656-.805-.812l-.578-.165v-.812l2.281.078 2.109-.078v.812l-.406.32c-.812.328-.898.406-1.219 1.062l-3.57 8.359c-2.117 4.797-4.312 5.203-5.852 5.203-.976.001-1.71-.249-2.351-.655z'
-    />
-  </Svg>
-)
-
-const Etsy = styled(EtsyBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #f45800;
-      }
-    `
-  })};
-`
-
-const EvenBriteBase = props => (
-  <Svg height={35} viewBox='0 0 26.872 4.549' {...props}>
-    <path
-      className='logo'
-      strokeWidth='.353'
-      d='M20.797.132c.129 0 .237.047.326.138a.464.464 0 0 1-.326.794.458.458 0 0 1-.33-.132.438.438 0 0 1-.133-.326c0-.135.044-.248.133-.336a.44.44 0 0 1 .33-.138zm-5.925.094a.786.786 0 0 1 .32 0c.072.015.14.045.203.088a.451.451 0 0 1 .155.204.853.853 0 0 1 .066.347v.656c0 .022.008.028.027.017.169-.087.375-.132.618-.132.41 0 .75.13 1.02.39.267.263.402.611.402 1.048 0 .478-.147.862-.441 1.147-.295.284-.715.424-1.262.424a3.43 3.43 0 0 1-1.037-.17c-.066-.014-.099-.063-.099-.144V.271c0-.01.008-.026.028-.045zM.772.518H2.53c.029 0 .044.015.044.044v.662c0 .026-.015.038-.044.038H.976c-.027 0-.039.015-.039.044v.717c.007.03.02.044.039.044h1.378c.029 0 .044.015.044.044v.634c0 .03-.015.044-.044.044H.976c-.02 0-.032.015-.039.044v.722c0 .03.012.044.039.044h1.61c.017 0 .03.013.038.034a.59.59 0 0 1 0 .314.536.536 0 0 1-.083.193.458.458 0 0 1-.198.149.825.825 0 0 1-.337.06H.176c-.03 0-.044-.014-.044-.044V1.158a.628.628 0 0 1 .64-.64zm11.84.215h.689c.025 0 .038.014.038.039v.656c0 .03.014.044.045.044h.733c.029 0 .044.015.044.044v.562c0 .026-.015.039-.044.039h-.733c-.03 0-.045.013-.045.038V3.27c0 .29.13.435.392.435.139 0 .288-.038.446-.116.011-.003.018-.003.022 0l.017.011.01.006a.44.44 0 0 1 .083.204.511.511 0 0 1-.05.287c-.114.21-.362.314-.743.314-.276 0-.505-.073-.684-.22-.176-.15-.264-.387-.264-.711V2.338a.219.219 0 0 0-.066-.155.219.219 0 0 0-.155-.066h-.17c-.031 0-.045-.012-.045-.038v-.563c0-.029.014-.044.044-.044h.353c.025 0 .039-.014.039-.044V.772c0-.025.013-.038.044-.038zm9.492 0h.694c.022 0 .033.014.033.039v.656c.007.03.022.044.044.044h.74c.024 0 .038.015.038.044v.562c0 .026-.014.039-.039.039h-.739c-.022 0-.037.013-.044.038V3.27c0 .29.13.435.392.435.13 0 .282-.038.452-.116l.044.017a.357.357 0 0 1 .066.11.478.478 0 0 1 .022.182.37.37 0 0 1-.055.199c-.108.209-.356.314-.744.314-.276 0-.503-.073-.678-.22-.177-.151-.265-.388-.265-.712v-1.14A.219.219 0 0 0 22 2.182a.224.224 0 0 0-.16-.066h-.165c-.026 0-.039-.013-.039-.039v-.562c0-.03.013-.044.039-.044h.347c.029 0 .044-.014.044-.044V.772c0-.025.013-.039.039-.039zm-16.928.739h.711c.043 0 .054.029.033.082L4.873 4.288c-.015.042-.035.061-.055.061h-.844c-.026 0-.045-.02-.055-.06L2.883 1.587a.049.049 0 0 1 .01-.056.515.515 0 0 1 .166-.082.827.827 0 0 1 .243-.044.422.422 0 0 1 .275.104.74.74 0 0 1 .226.342l.585 1.67c.003.011.008.017.016.017.01 0 .018-.006.022-.017l.695-1.995c.01-.036.029-.055.055-.055zm2.216-.066c.834 0 1.251.497 1.251 1.488a.228.228 0 0 1-.072.165.227.227 0 0 1-.165.066H6.791c-.026 0-.039.014-.039.039a.6.6 0 0 0 .287.435c.164.104.349.155.551.155.29 0 .579-.064.866-.193.018-.01.035-.007.055.01a.623.623 0 0 1 .077.37.377.377 0 0 1-.138.243.828.828 0 0 1-.43.182 2.453 2.453 0 0 1-.121.022l-.177.016c-.07.008-.142.011-.215.011-.467 0-.836-.132-1.107-.397-.273-.263-.408-.623-.408-1.075 0-.465.13-.839.391-1.119.26-.278.597-.418 1.009-.418zm3.114 0c.396 0 .7.09.91.264.212.174.32.465.32.871v1.814c0 .019-.01.03-.028.033a.372.372 0 0 1-.133.027.628.628 0 0 1-.397-.105.478.478 0 0 1-.16-.209.871.871 0 0 1-.06-.347V2.596c0-.326-.17-.49-.507-.49-.17 0-.335.023-.496.066-.033.008-.05.03-.05.066v2.067c0 .03-.013.044-.038.044h-.69c-.03 0-.043-.014-.043-.044V1.703c0-.029.013-.05.044-.06a3.87 3.87 0 0 1 1.328-.237zm8.781 0c.197 0 .353.034.463.099.113.066.18.15.198.248.018.096.024.18.017.253a.565.565 0 0 1-.05.188c-.01.022-.03.03-.055.022a1.644 1.644 0 0 0-.55-.077c-.13 0-.252.02-.365.055-.019.008-.027.02-.027.038v2.073c0 .03-.016.044-.044.044h-.69c-.03 0-.044-.014-.044-.044v-2.59c0-.01.003-.02.011-.028l.017-.022a.19.19 0 0 1 .039-.028l.022-.011a2.66 2.66 0 0 1 1.058-.22zm6.207 0c.83 0 1.245.497 1.245 1.488a.227.227 0 0 1-.072.165.219.219 0 0 1-.16.066h-1.614c-.03 0-.045.014-.045.039.032.187.13.333.293.435a.989.989 0 0 0 .545.155c.29 0 .58-.064.866-.193.018-.01.036-.007.055.01a.388.388 0 0 1 .066.15.74.74 0 0 1 .01.22c0 .086-.047.167-.132.243a.881.881 0 0 1-.363.17c-.01 0-.03.005-.067.012a2.453 2.453 0 0 1-.12.022l-.177.016a2.004 2.004 0 0 1-.215.011c-.467 0-.84-.132-1.114-.397-.272-.263-.408-.623-.408-1.075 0-.465.13-.839.392-1.119.263-.278.602-.418 1.014-.418zm-5.038.066h.689c.029 0 .044.015.044.044v2.839c0 .019-.01.03-.028.033a.428.428 0 0 1-.138.027.698.698 0 0 1-.193-.016.512.512 0 0 1-.198-.088.504.504 0 0 1-.154-.21.857.857 0 0 1-.061-.342V1.516c0-.029.012-.044.039-.044zm-13.08.573a.551.551 0 0 0-.398.154.835.835 0 0 0-.22.42c-.004.014-.003.025.006.032.007.004.02.006.039.006h1.074c.025 0 .039-.012.039-.038a.651.651 0 0 0-.149-.414c-.088-.106-.219-.16-.391-.16zm18.101 0a.558.558 0 0 0-.402.154.81.81 0 0 0-.215.42c-.004.014-.003.025.006.032.007.004.02.006.038.006h1.07c.025 0 .038-.012.038-.038a.651.651 0 0 0-.148-.414c-.089-.106-.218-.16-.386-.16zm-9.343.066a.913.913 0 0 0-.502.138c-.01.004-.016.015-.016.033v1.323c0 .052.022.087.066.105.142.03.254.044.33.044.261 0 .472-.073.635-.22.16-.15.242-.373.242-.668a.736.736 0 0 0-.215-.545.745.745 0 0 0-.54-.21zm0 0'
-    />
-  </Svg>
-)
-
-const EventBrite = styled(EvenBriteBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #f5662e;
-      }
-    `
-  })};
-`
-
-const TimeBase = props => (
-  <Svg height={64} viewBox='0 0 298 92.321' {...props}>
-    <path
-      className='logo'
-      d='M142.27.951s-3.33.07-5.26.037c-.08.81-1.01 8.15-1.81 14.053-.92 5.463-8.21 45.988-10.68 58.12-.65 3.662-2.66 14.01-2.85 15.044-2.32.277-5.06.533-6.85.602-.24.008-.64.086-.83.086v2.101c.22-.009 1.38.08 2.2.045 1.97-.063 7.56-.328 8.62-.328 4.02 0 6.7.266 10.42.34.05 0 1.07-.004 1.52-.06v-2.124c-.31 0-.72-.01-.96-.01-1.05.033-4.92-.264-6.88-.53.23-1.389 3.65-22.377 5.29-30.77 3.12-15.535 5.28-26.545 5.28-26.545l.58 1.367 24.65 58.752 4.5.023s2.57-6.362 5.62-13.498c.08-.169.3-.68.61-1.41 3.2-7.41 19.06-45.017 19.06-45.017s.15.677.4 1.888c1.51 7.371 6.62 33.935 7.71 40.344.51 3.297.99 6.641 1.27 8.6.41 2.774.68 5.49.68 6.27-.5.022-5.52.476-7.42.558-.27-.009-.78.016-.81.021v2.137c.68 0 1.34.03 2 .015 3.88-.074 7.61-.378 11.49-.376 4.36.004 11.43.385 15.96.414.51 0 1.42-.045 1.42-.045v-2.159c-.26-.002-1.4-.058-1.82-.09-2.03-.16-5.45-.477-6.32-.535 0 0-.56-3.103-1.3-6.687-2.96-14.337-11.39-57.916-12.68-68.639-.44-3.685-1.69-11.96-1.69-11.96l-5.07-.002-4.95 14.207c-1.61 4.63-3.8 10.346-5.47 14.729-8.24 21.656-17.48 41.836-17.48 41.836s-19.64-46.862-20.68-49.572c-1.05-2.711-7.47-21.231-7.47-21.231zM1.09.971l.043 4.953-.084 12.062 1.307-.037.136-10.728s7.166-.062 8.092-.061l13.361.006 8.623.043.014 74.607.002 6.444-.182.025c-1.148.089-5.565.359-7.359.424-.402.013-.659.023-.703.018l.008 2.183a38.4 38.4 0 0 0 1.312.033c4.244-.055 9.987-.424 14.06-.42 4.136.002 9.958.404 13.561.446.357.002.78-.033 1.123-.033v-2.047l-.447.004c-1.148-.005-7.497-.414-8.154-.5a2244.97 2244.97 0 0 1-.145-5.94L45.69 7.2l6.565-.018 17.13.012 5.716.078c.038.409.327 5.843.322 5.99-.013.396.214 4.444.205 5.44l1.355.016-.097-12.2.062-5.488c-5.034.137-17.863.31-37.773.299C22.1 1.32 5.644 1.176 1.09.97zm79.746.02l.021 2.136c.676.037 4.65.388 5.616.436.771.094 1.847.19 2.617.265 0 .684.012 5.405.023 7.268.129 17.659.019 56.687-.033 70.496l-.027 6.566c-.487.146-1.583.201-2.768.281-1.013.069-4.643.374-5.453.391v2.137c.336 0 .952.071 1.303.068 4.109.001 9.822-.509 14.031-.506 4.254.003 9.934.521 13.284.522l-.01.002c.29.004.98-.072 1.27-.071v-2.136h-.29c-.13-.028-4.45-.21-5.5-.319-.55-.035-2.93-.241-2.96-.32-.01-.042-.02-.681-.02-1.746.01-1.011.01-2.692.01-4.514.01-13.929.05-55.298.06-71.111 0-3.056.01-5.54.01-6.871.93-.059 4.73-.456 6.09-.576.06-.006 1.86-.11 2.6-.155V1.008c-4.65-.002-10.803.362-15.341.36C90.623 1.363 84.776.991 80.836.99zm215.2.06l-1.4.004c-3.11.013-11.07.072-20.38.123-5.72.033-10.19.079-16.08.076-4.13.048-9.91.208-13.48.201-5.7-.01-10.35-.36-12.88-.371-1.2-.005-1.87.04-2.2.02.02.03.02 2.109.02 2.109.22.038.66.087 1.07.078 1.43.096 4.11.313 5.47.492.27.05.73.126 2.01.268-.01.473.06 5.827.05 9.043 0 2.98-.03 69.47-.03 69.47 0 .108-.15 5.347-.15 5.77-2.26.387-6.6.625-8.08.631h-.44v2.115c.21-.001 1.18.037 1.73.037 2.39-.001 8.55-.381 14.84-.385 4.41-.004 9.26.161 14.17.231 6.05.084 12.17.088 17.67.166 9.07.128 16.42.139 18.81.142 0 0 .29-16.823.29-18.179l-1.31.015c0 1.431-.53 6.76-.55 6.993-.17 2.006-.35 4.108-.49 5.144 0 0-.64.051-.95.045-1.49-.028-6.25.145-6.35.145l-21.23.009-14.91-.25-.1-4.136-.01-32.112c2.66-.09 6.72-.231 9.19-.226 2.1.004 7.05-.114 9.82-.108 6.73.012 12.71.055 13.19.069 1.11.032 3.17.1 3.95.148.03.283.12.725.13.85.4 2.459.62 4.162.89 6.576.04.217.06.427.09.625h1.53s.04-.863.02-1.4c-.03-1.893-.15-5.669-.15-9.854-.01-2.901.16-7.336.18-9.469.01-.577-.06-.987-.06-1.152l-1.4.027c-.01.461-.12 1.141-.12 1.141-.12 1.59-.57 3.622-.93 5.766-.1.267-.17.64-.17.64-.16.054-4.55.288-4.94.285-.15-.001-6.54.178-6.65.178l-24.56-.092.04-35.41s3.12-.11 3.99-.105c3.54 0 9.3-.032 15.17-.024 8.59.01 16.93.064 16.95.076.05.025 2.59.06 2.64.104.06.048 3.24.143 4.02.168 0 0 .06.263.1.595.35 3.734.64 8.168.77 9.739l.01.629 1.31-.004c0-2.299-.12-9.34-.14-13.819-.01-1.439.02-2.61.02-3.277v-.57z'
-    />
-  </Svg>
-)
-
-const Time = styled(TimeBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #e90606;
-      }
-    `
-  })};
-`
-
-const BBC = props => (
-  <Svg height={59.5} version={1} viewBox='0 0 300 86.219' {...props}>
-    <path d='M162.5 55.573c0 9.26-11.521 8.71-11.521 8.71h-11.523V47.55h11.523c11.865-.069 11.522 8.024 11.522 8.024m-23.045-33.538h8.78c9.121.48 8.778 7.27 8.778 7.27 0 8.436-10.082 8.573-10.082 8.573h-7.476V22.035zm20.92 19.135s7.886-3.36 7.818-12.277c0 0 1.199-14.608-18.159-16.391h-21.484v61.246h24.623s20.575.069 20.575-17.284c0 0 .48-11.796-13.374-15.294M104.784.5h90.432v85.22h-90.432V.5zM58.217 55.573c0 9.26-11.522 8.71-11.522 8.71H35.172V47.55h11.523c11.865-.069 11.522 8.024 11.522 8.024M35.172 22.035h8.78c9.121.48 8.778 7.27 8.778 7.27 0 8.436-10.082 8.573-10.082 8.573h-7.476V22.035zm20.92 19.135s7.886-3.36 7.818-12.277c0 0 1.199-14.608-18.159-16.391H24.267v61.246H48.89s20.575.069 20.575-17.284c0 0 .48-11.796-13.374-15.294M.5.5h90.432v85.22H.5V.5zm280.55 16.392V28.21s-11.042-6.79-23.25-6.928c0 0-22.77-.444-23.799 21.829 0 0-.823 20.489 23.525 21.655 0 0 10.22 1.234 24.073-7.613V68.88s-18.586 11.042-40.122 2.537c0 0-18.107-6.617-18.793-28.307 0 0-.755-22.308 23.388-29.716 0 0 6.447-2.469 18.038-1.372 0 0 6.927.687 16.94 4.87M209.068 85.72H299.5V.5h-90.432v85.22z' />
-  </Svg>
-)
-
-const NasaBase = props => (
-  <Svg
-    version='1.1'
-    xmlnsXlink='http://www.w3.org/1999/xlink'
-    x={0}
-    y={0}
-    height={57}
-    viewBox='0 0 145 40'
-    {...props}
-  >
-    <defs xmlns='http://www.w3.org/2000/svg'>
-      <clipPath id='a'>
-        <rect x={0} y={0} width={145} height={39} />
-      </clipPath>
-    </defs>
-    <path
-      className='logo'
-      stroke='black'
-      d='M5 39V10a6 6 0 0 1 6-6 5.993 5.993 0 0 1 5.644 3.988l6.712 24.024A5.993 5.993 0 0 0 29 36a6 6 0 0 0 6-6V1'
-      fillOpacity={0}
-      strokeWidth={8}
-    />
-    <path
-      d='M40.5 41.5L51.356 7.988A5.992 5.992 0 0 1 57 4a5.992 5.992 0 0 1 5.644 3.988L73.5 41.5'
-      fillOpacity={0}
-      className='logo'
-      stroke='black'
-      strokeWidth={8}
-      clipPath='url(#a)'
-    />
-    <path
-      d='M104 5H84.494a7.503 7.503 0 0 0-7.504 7.502 7.505 7.505 0 0 0 7.51 7.502h10A7.5 7.5 0 1 1 94.5 35L74 34.992'
-      fillOpacity={0}
-      className='logo'
-      stroke='black'
-      strokeWidth={8}
-      strokeMiterlimit={0}
-    />
-    <path
-      d='M106.625 41.5L117.5 7.988a6.002 6.002 0 0 1 11.308 0L139.683 41.5'
-      fillOpacity={0}
-      className='logo'
-      stroke='black'
-      clipPath='url(#a)'
-      strokeWidth={8}
-    />
-  </Svg>
-)
-
-const Nasa = styled(NasaBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        stroke: #f00;
-      }
-    `
-  })};
-`
-
-const TEDBase = props => (
-  <Svg height={76} viewBox='0 0 324 119' {...props}>
-    <path
-      className='logo'
-      d='M32.678 30.831H0V.85h101.56v29.981H68.875V118.3H32.678V30.831zM107.18.85h98.75v29.981h-62.55v14.462h62.55v28.212h-62.55v14.814h62.55V118.3h-98.75V.85zm104.72 0h59.39C310.3.85 324 29.773 324 59.401c0 35.975-18.98 58.899-59.74 58.899H211.9V.85zm36.2 87.469h14.05c22.49 0 25.66-18.337 25.66-29.27 0-7.41-2.46-27.865-28.47-27.865h-11.6l.36 57.135z'
-    />
-  </Svg>
-)
-
-const TED = styled(TEDBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #e62b1e;
-      }
-    `
-  })};
-`
-
-const BuzzFeedBase = props => (
-  <Svg height={36.5} viewBox='0 0 853.3 150.1' {...props}>
-    <path
-      className='logo'
-      d='M0 147.3V0h65.7c28 0 42.4 17.9 42.4 37.5 0 18.3-11.5 30.9-25.4 33.8 15.9 2.4 28.3 17.7 28.3 36 0 22.3-14.6 40-42.8 40H0zM76.1 43.1c0-9.1-6.6-15.5-17.2-15.5H31.4v30.9h27.5c10.6 0 17.2-5.9 17.2-15.4zm2.9 59.8c0-9.3-6.6-16.8-18.8-16.8H31.4v33.6h28.9c11.6 0 18.7-6.2 18.7-16.8zm40.6 1.7v-64h28.2v63.1c0 12.7 7.7 21.9 22.7 21.9 14.8 0 22.6-9.2 22.6-21.9V40.6h28.1v63.9c0 26.8-16.1 45.6-50.6 45.6s-51-19-51-45.5zm110.9 42.7v-21l48.1-61.2h-48.1V40.6h86.3v20.1l-49 62.1h49.9v24.5h-87.2zm96.4 0v-21L375 65.1h-48.1V40.6h86.3v20.1l-49 62.1h49.9v24.5h-87.2zm97.7 0V0h102.2v27.6H456v30.9h53.3v27.6H456v61.2h-31.4zm92.6-53.4c0-30.9 23-55.9 55.2-55.9 32 0 53.4 23.9 53.4 58.5v6.6h-79.3c2 13 12.6 23.9 30.7 23.9 9.1 0 21.4-3.8 28.3-10.4l12.6 18.6c-10.6 9.7-27.4 14.8-43.9 14.8-32.5 0-57-21.9-57-56.1zM572.4 61c-17.4 0-25 12.1-26.1 22.7h52.6c-.9-10.2-8-22.7-26.5-22.7zm61 32.9c0-30.9 23-55.9 55.2-55.9 32 0 53.4 23.9 53.4 58.5v6.6h-79.3c2 13 12.6 23.9 30.7 23.9 9.1 0 21.4-3.8 28.3-10.4l12.6 18.6c-10.6 9.7-27.4 14.8-43.9 14.8-32.5 0-57-21.9-57-56.1zM688.6 61c-17.4 0-25 12.1-26.1 22.7h52.6c-.9-10.2-8-22.7-26.5-22.7zm63.7 86.5V40.4h43c33.7 0 58 20.2 58 53.5 0 33.2-24.2 53.6-57.8 53.6h-43.2zM826 93.9c0-16.6-10-30.4-29.6-30.4H779v60.9h17.4c18.7 0 29.6-14.6 29.6-30.5z'
-    />
-  </Svg>
-)
-
-const BuzzFeed = styled(BuzzFeedBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #ee3322;
-      }
-    `
-  })};
-`
-
-const FoxBase = props => (
-  <Svg version={1} height={89.4} viewBox='0.4 -4 1829.088 789.32' {...props}>
-    <path
-      className='logo'
-      d='M1829.488 783.9h-250.1c-2.3 0-6.5-10.1-7.7-12.2-34.4-62.1-68.8-124.5-103-186.5-36.6 66-72.2 133.1-108.7 199.2-81.9.2-163.1-.4-244.3-.2 9.2-17.2 19-34 28.8-50.9 68.3-118 135.6-237.6 204-355.7-72.7-127-146.5-253-219.6-379.7-.2 0-.7-.8-.2-.7h251.3c30.9 56.2 61 113.1 91.4 169.7 31.6-56.1 62.1-113.3 93.4-169.7h242.9c.2 1-.5 2.1-1 2.9-70.9 124.6-143.5 249.6-214.3 374.4 78.7 136.6 158.9 273.1 237.1 409.4z'
-    />
-    <path
-      className='logo'
-      d='M1222.289 302.5c6.4 26 10.1 54.6 10.1 84.7 0 93.9-27.2 163.5-65 220.1-6.4 9.6-12.9 18.8-19.9 27.6-20.5 26-45.3 49.7-73 69.4-18.6 13.2-38 25.2-59.5 35.3-42.8 20.1-94.5 35.8-154.1 35.8-93.2 0-162-28.4-217.7-67.4-55.7-39.1-99.1-90.8-128.6-156-19.8-43.7-32.8-94.1-34.3-155.5-.7-30.4 3.1-60.4 8.6-86.4 14-65.6 41.5-117 76.3-161.3 34.8-44.2 78.9-78.8 132.2-104.9 20.9-10.2 44.3-18.9 69.8-25.2 25.5-6.3 53.7-10.6 84.2-10.6h.5c29.8-1.4 61.4 3.4 86.2 8.9 25.8 5.7 49.7 14.4 71 24 66 29.8 115.2 74.4 154.8 130.1 26.1 36.5 45.7 80.3 58.4 131.4zm-300 260.6c2.4-9.6 1.7-21.9 1.7-34.6v-289c0-12.4-1.4-22-5-31-3.3-8.2-8.4-15.7-14.4-21.8-11.7-11.9-29.3-22.6-51.4-20.9-10.7.8-19.6 4.2-27.4 9.1-14.5 9.2-25.5 23.3-30 42-2.3 9.5-2.2 21.6-2.2 34.1v289c0 6.2-.3 12.2.5 17.5 2.1 15.2 8.9 27.4 17.8 36.5 11.5 11.7 29.3 21.5 52.3 19 29.7-3.2 51.5-22.9 58.1-49.9zm-431.8-351.3h-265.7c-1.1 33-.2 69.5-.5 102.7 71.8.6 144.7.1 217 .2V530c-72.8.2-146.2-.3-218.6.2.1 85.4.5 169.2.5 254.6 0 .3-.1.4-.2.5H1.489c-.4.1-.4-.2-.7-.2-.5-262.2-.5-526.1-.1-788.4 0-.3 0-.7.2-.7h474.5c.3 0 .4 0 .5.2 5 71.8 9.8 143.7 14.6 215.6z'
-    />
-  </Svg>
-)
-
-const Fox = styled(FoxBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo {
-        fill: #002885;
-      }
-    `
-  })};
-`
-
-const TechCrunchBase = props => (
-  <Svg height={141.5} viewBox='-3.5 -2.607 290 198.312' {...props}>
-    <path
-      className='logo'
-      d='M141.5-2.607h-145v48h48v99h49v-99h48zm48 0h97v48h-97zm0 96.449V45.393h-48v99h145v-50.48zm-173 70.551v30h-8v-30h-12v-6h30v6h-10zm32.47 18H32.333c.264 4 3.073 5.834 5.723 5.834 2.014 0 3.708-.447 4.769-1.669l5.192 4.491c-2.225 2.86-6.253 4.578-9.961 4.578-8.001 0-13.088-5.632-13.088-14.428 0-8.531 5.351-14.501 12.4-14.501 7.419 0 11.709 5.989 11.709 13.938 0 1.065-.053 1.757-.107 1.757zm-11.973-8.982c-2.65 0-4.292.982-4.557 4.982h9.538c-.212-4-1.696-4.982-4.981-4.982zm29.512 22.294c-7.26 0-13.034-5.935-13.034-14.466s5.617-14.519 13.352-14.519c5.299 0 8.69 2.597 11.022 6.148l-4.821 4.874c-1.643-2.332-3.391-3.975-6.358-3.975-3.498 0-5.776 2.914-5.776 7.47 0 4.453 2.278 7.421 5.776 7.421 2.86 0 4.662-1.271 6.571-3.55l4.662 4.397c-2.704 3.604-5.884 6.2-11.394 6.2zm32.991-1.312v-14.32c0-4.979-.953-6.463-3.974-6.463s-4.026 1.485-4.026 6.307v14.478h-8v-34.241l8-3.659v12.667c0-1.325 3.205-2.437 6.119-2.437 7.101 0 9.881 4.874 9.881 12.716v14.954l-8-.002z'
-    />
-    <path
-      className='logo'
-      d='M136.715 168.84c-1.325-2.542-3.444-4.027-6.781-4.027-5.564 0-8.742 4.876-8.742 11.763 0 6.676 3.232 11.763 8.795 11.763 3.341 0 5.301-1.43 7.26-4.715l6.253 3.708c-2.542 5.246-6.996 8.372-13.723 8.372-9.697 0-16.216-7.421-16.216-19.13 0-11.286 6.781-19.128 16.478-19.128 6.996 0 11.235 3.391 13.564 8.211l-6.888 3.183zm27.236 6.994c-1.166-1.061-2.346-1.802-3.988-1.802-2.859 0-4.463 1.222-4.463 6.04v14.32h-8v-28h8v2.767c2-1.325 2.876-2.437 5.473-2.437 2.383 0 4.057.952 5.01 1.96l-2.032 7.152zm23.549 18.31v-.876c-2 1.324-3.205 2.437-6.119 2.437-7.102 0-9.881-4.876-9.881-12.716v-16.596h6v15.958c0 4.981 1.953 6.465 4.975 6.465 3.016 0 5.025-1.486 5.025-6.308v-16.116h8v28h-8.314l.314-.248zm29 .249v-14.32c0-4.979-.453-6.463-3.475-6.463-3.016 0-3.525 1.485-3.525 6.307v14.478h-8v-28h8v2.767c0-1.325 2.705-2.437 5.619-2.437 7.102 0 9.381 4.874 9.381 12.716v14.954l-8-.002zm27.352 1.312c-7.26 0-13.033-5.935-13.033-14.466s5.617-14.519 13.352-14.519c5.301 0 8.691 2.597 11.021 6.148l-4.818 4.874c-1.645-2.332-3.391-3.975-6.357-3.975-3.5 0-5.779 2.914-5.779 7.47 0 4.453 2.279 7.421 5.779 7.421 2.857 0 4.662-1.271 6.57-3.55l4.662 4.397c-2.709 3.604-5.886 6.2-11.397 6.2zm32.648-.582v-15.05c0-4.979-.953-6.463-3.975-6.463-3.016 0-4.025 1.485-4.025 6.307v14.478h-8v-34.241l8-3.659v12.667c0-1.325 3.205-2.437 6.119-2.437 7.102 0 9.881 4.874 9.881 12.716v14.954h-7.812l-.188.728z'
-    />
-  </Svg>
-)
-
-const TechCrunch = styled(TechCrunchBase)`
-  ${style};
-  ${createCssState({
-    selector: '&:hover:not([disabled])',
-    state: 'hover',
-    css: css`
-      .logo:nth-child(1) {
-        fill: #0a9e01;
-      }
-      .logo:nth-child(2) {
-        fill: #343c34;
-      }
-    `
-  })};
-`
-
-const logos = [
-  Amazon,
-  Apple,
-  BBC,
-  BuzzFeed,
-  CNN,
-  Etsy,
-  EventBrite,
-  Facebook,
-  Flickr,
-  Fox,
-  GitHub,
-  Gizmodo,
-  Instagram,
-  Medium,
-  Nasa,
-  NYTimes,
-  Reddit,
-  SoundCloud,
-  TechCrunch,
-  TED,
-  TheGuardian,
-  TheWashingtonPost,
-  Time,
-  Twitter,
-  Vimeo,
-  Wikipedia,
-  YouTube
-]
+const logos = Object.entries(Logos)
 
 const states = ['normal', 'hover']
 
@@ -775,11 +16,12 @@ function LogoStories () {
           as='article'
           flexDirection='column'
           alignItems='baseline'
+          key={state}
         >
           <Subhead>{state}</Subhead>
-          {logos.map((LogoComponent, index) => (
-            <Flex py={3} alignItems='center'>
-              <LogoComponent state={state} key={`logo_${index}`} />
+          {logos.map(([logoName, LogoComponent], index) => (
+            <Flex py={3} alignItems='center' key={`${logoName}_${state}`}>
+              <LogoComponent state={state} />
             </Flex>
           ))}
         </Flex>
