+++ src/components/logos/Microlink.js
@@ -1,11 +1,13 @@
 import { Image } from 'components/elements'
 import styled from 'styled-components'
 
+import logo from '../../../static/logo.svg'
+
 const Microlink = styled(Image)([])
 
 Microlink.defaultProps = {
   ...Image.defaultProps,
-  src: '/logo.svg',
+  src: logo,
   style: { height: 'auto' }
 }
